% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEOmics.R
\name{annoSncList}
\alias{annoSncList}
\title{Annotate the type of small RNA in the differential expression results}
\usage{
annoSncList(deg, annoList)
}
\arguments{
\item{deg}{A data frame containing the differential expression analysis.
Format consistent with the output of exportDE()}

\item{annoList}{A list of annotations for small RNA.}
}
\value{
A data frame containing the differential expression analysis with
        the type of small RNA annotated.
}
\description{
This function annotates the type of small RNA in the
             differential expression results.
}
\references{
\insertRef{dplyr}{IntegraTRN}
}
\keyword{internal}
