% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processPeaks.R
\name{annotateATACPeaksMotif}
\alias{annotateATACPeaksMotif}
\title{Performing peak annotation and motif enrichment analysis}
\usage{
annotateATACPeaksMotif(
  objMOList,
  tssRegion = c(-3000, 3000),
  TxDb,
  annoDb = "org.Hs.eg.db",
  bsgenome,
  pwmL,
  fixedWidth = 500
)
}
\arguments{
\item{objMOList}{An object of class MOList}

\item{tssRegion}{A numeric vector of length 2, specifying the region around
the TSS to annotate with (default: +/- 3000). The first and
second elements of the vector specify the base pairs
upstream and downstream of the TSS, respectively.}

\item{TxDb}{The TxDb object to use for annotation, using one of the TxDb
packages. See examples in how to generate the TxDb object.}

\item{annoDb}{The annotation database to use for annotation, using one of
 the org.*.eg.db packages. Must be a valid string for the name
 of the package. In addition, before calling the function, the
 user must load the corresponding annotation package using
'library(org.*.eg.db)'. For example, if the user wants to use
the org.Hs.eg.db package for human samples, the user must first
load the package using 'library(org.Hs.eg.db)' and then supply
the string "org.Hs.eg.db" to the annoDb argument. For details,
see the vignette of the ChIPseeker package for
custom annotation databases.}

\item{bsgenome}{The BSgenome object that contains the genome sequence.
Should be a valid BSgenome object generated using the
one of the BSgenome packages. The genome sequence should
correspond to the genome version used for the TxDb object.
Read the Bioconductor vignette for the BSgenome package for
more details.}

\item{pwmL}{A PWMatrixList object containing the position-weight matrices
(PWMs) to use for motif enrichment analysis. The PWMs can be
generated using the getMatrixSet() function from the TFBSTools
package. See the vignette of the TFBSTools package for more
details.}

\item{fixedWidth}{The fixed width to adjust the peak size for motif if
necessary (default: 500)}
}
\value{
An object of class MOList, with the ATACseq peaks annotated with
        genomic features and motif enrichment analysis results appended.
}
\description{
This function performs peak annotation and motif enrichment
             analysis on the ATACseq peaks. The peak annotation is performed
             using the ChIPseeker package, and the motif enrichment analysis
             is performed using the monaLisa package.
}
\details{
The annotation is performed using the TxDb object and annotation
         database specified by the user. Depending on the style of genomic
         coordinate representation (with or without the "chr" prefix for the
         chromosome name, or the use of "chrM", "M", "MT" for mitochondrial
         DNA), the TxDb object may need to be adjusted to match the style
         used in the peak file. See the vignette of the ChIPseeker package
         for more details. By default, we recommend using the UCSC style
         genomic coordinates, and the use of the
         TxDb.Hsapiens.UCSC.hg38.knownGene and org.Hs.eg.db packages for
         human samples, and the TxDb.Mmusculus.UCSC.mm10.knownGene and
         org.Mm.eg.db packages for mouse samples.
}
\examples{
\dontrun{
# Generate position-weight matrices (PWMs) from the JASPAR database

# This is an example of how to generate the PWMs from the JASPAR database
# Alternative databases can be used, as long as a PWMatrixList object is
# generated

# This example requires users to have the JASPAR2022 package installed,
# which is only available on Bioconductor 3.18 or later
# In addition, the TFBSTools package is required

# Generate the PWMs
pwmL <- TFBSTools::getMatrixSet(JASPAR2022::JASPAR2022,
  opts = list(
    matrixtype = "PWM",
    tax_group = "vertebrates"
  )
)
# Alternatively, the package has provided a pre-generated PWM list
data("jasparVertebratePWM")

# Make use of the package's provided example data
data("expMOList")

# Annotate the ATACseq peaks with genomic features and perform motif
# enrichment analysis
expMOList <- annotateATACPeaksMotif(expMOList,
  tssRegion = c(-3000, 3000),
  TxDb = TxDb.Hsapiens.UCSC.hg38.knownGene::TxDb.Hsapiens.UCSC.hg38.knownGene,
  annoDb = "org.Hs.eg.db",
  bsgenome = BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38,
  pwmL = jasparVertebratePWM,
  fixedWidth = 500
)
}

}
