% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processPeaks.R
\name{annotatePeaks}
\alias{annotatePeaks}
\title{Annotate peaks with genomic features}
\usage{
annotatePeaks(
  objMOList,
  tssRegion = c(-3000, 3000),
  TxDb,
  annoDb = "org.Hs.eg.db"
)
}
\arguments{
\item{objMOList}{An object of class MOList}

\item{tssRegion}{The region around the TSS to annotate with
(default: +/- 3000)}

\item{annoDb}{The annotation database to use for annotation, using one of
the org.*.eg.db packages. Must be a valid string for the name
of the package. For details, see the vignette of the ChIPseeker
package for custom annotation databases.}

\item{TxDB}{The TxDb object to use for annotation, using one of the TxDb
packages}
}
\value{
A csAnno object containing the annotated peaks, as defined by the
        ChIPseeker package. In particular, the original "Condition"
        annotation is preserved.
}
\description{
This function annotates the ATACseq peaks with genomic features
             using the ChIPseeker package.
}
\details{
The annotation is performed using the TxDb object and annotation
         database specified by the user. Depending on the style of genomic
         coordinate representation (with or without the "chr" prefix for the
         chromosome name, or the use of "chrM", "M", "MT" for mitochondrial
         DNA), the TxDb object may need to be adjusted to match the style
         used in the peak file. See the vignette of the ChIPseeker package
         for more details. By default, we recommend using the UCSC style
         genomic coordinates, and the use of the
         TxDb.Hsapiens.UCSC.hg38.knownGene and org.Hs.eg.db packages for
         human samples, and the TxDb.Mmusculus.UCSC.mm10.knownGene and
         org.Mm.eg.db packages for mouse samples.
}
\references{
\insertRef{yu2015chipseeker}{IntegraTRN}

\insertRef{lawrence2013software}{IntegraTRN}
}
\keyword{internal}
