% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processSmallRNA.R
\name{annotateSmallRNA}
\alias{annotateSmallRNA}
\title{Annotate small RNA transcripts to respective categories}
\usage{
annotateSmallRNA(objMOList, anno = "human")
}
\arguments{
\item{objMOList}{A object of class MOList}
}
\value{
A MOList object with the annotation information added to the
       annoSncRNA element.
}
\description{
This function annotates small RNA transcripts to their
             respective categories. The categories are defined as follows:
             1. miRNA
             2. piRNA
             3. tRNA
             4. circRNA
             5. snRNA
             6. snoRNA
}
\examples{
# Example 1: Adding user-defined annotation
# Generate some sample RNAseq and smallRNAseq data
rnaMatrix <- matrix(sample(0:100, 1000, replace = TRUE), nrow = 100)
rownames(rnaMatrix) <- paste0("gene_", seq_len(100))
rnaGroupBy <- rep(c("group1", "group2"), each = 5)
smallRNAMatrix <- matrix(sample(0:100, 1000, replace = TRUE), nrow = 100)
rownames(smallRNAMatrix) <- paste0("transcript_", seq_len(100))
smallRNAGroupBy <- rep(c("group1", "group2"), each = 5)

# Create example MOList object
objMOList <- MOList(
  RNAseq = rnaMatrix, RNAGroupBy = rnaGroupBy,
  smallRNAseq = smallRNAMatrix, smallRNAGroupBy = smallRNAGroupBy
)

# Create example annotation
anno <- data.frame(
  transcript = paste0("transcript_", seq_len(120)),
  category = rep(c("miRNA", "piRNA", "tRNA", "circRNA", "snRNA", "snoRNA"),
    each = 20
  )
)

# Annotate small RNA transcripts
annotateSmallRNA(objMOList, anno)

# Example 2: Using the preprocessed annotation provided by the package
# This time we use the package provided example data
data(expMOList)

# Annotate small RNA transcripts
annotateSmallRNA(expMOList, anno = "human")

# or simply use the default parameter
annotateSmallRNA(expMOList)

}
\references{
\insertRef{li2020compsra}{IntegraTRN}

\insertRef{kozomara2010mirbase}{IntegraTRN}

\insertRef{chan2016gtrnadb}{IntegraTRN}

\insertRef{harrow2012gencode}{IntegraTRN}

\insertRef{glavzar2014circbase}{IntegraTRN}

\insertRef{sai2008pirnabank}{IntegraTRN}

\insertRef{zhang2014pirbase}{IntegraTRN}

\insertRef{rosenkranz2016pirna}{IntegraTRN}

\insertRef{dplyr}{IntegraTRN}
}
