% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEAKTag.R
\name{as.data.frame,PEAKTag-method}
\alias{as.data.frame,PEAKTag-method}
\title{Convert to a data frame}
\usage{
\S4method{as.data.frame}{PEAKTag}(x)
}
\arguments{
\item{x}{A PEAKTag object}
}
\value{
A data frame
}
\description{
Convert to a data frame
}
\examples{
\dontrun{
# Use the package's provided example data and position weight matrix
data("expMOList")
data("jasparVertebratePWM")

# Perform annotation and motif enrichment
expMOList <- annotateATACPeaksMotif(
  expMOList,
  TxDb = TxDb.Hsapiens.UCSC.hg38.knownGene::TxDb.Hsapiens.UCSC.hg38.knownGene,
  annoDb = "org.Hs.eg.db",
  bsgenome = BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38,
  pwmL = jasparVertebratePWM
)

# Once annotation and enrichment has been performed, the result becomes a
# PEAKTag object (the DEATAC element)

# Convert the object to a data frame
as.data.frame(expMOList$DEATAC)
}

}
