% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEAKTag.R
\name{asGRanges}
\alias{asGRanges}
\alias{asGRanges,PEAKTag-method}
\title{Convert to a GRanges object}
\usage{
asGRanges(x)
}
\arguments{
\item{x}{A PEAKTag object}
}
\value{
A GRanges object
}
\description{
Convert to a GRanges object
}
\examples{
\dontrun{
# Use the package's provided example data and position weight matrix
data("expMOList")
data("jasparVertebratePWM")

# Perform annotation and motif enrichment
expMOList <- annotateATACPeaksMotif(
  expMOList,
  TxDb = TxDb.Hsapiens.UCSC.hg38.knownGene::TxDb.Hsapiens.UCSC.hg38.knownGene,
  annoDb = "org.Hs.eg.db",
  bsgenome = BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38,
  pwmL = jasparVertebratePWM
)

# Once annotation and enrichment has been performed, the result becomes a
# PEAKTag object (the DEATAC element)

# Convert the object to a GRanges object
asGRanges(expMOList$DEATAC)
}

}
