% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOList.R
\name{ATACpeaks,MOList-method}
\alias{ATACpeaks,MOList-method}
\alias{ATACpeaks}
\title{Getter for the ATAC peaks data from the MOList object}
\usage{
\S4method{ATACpeaks}{MOList}(x)
}
\arguments{
\item{x}{An object of class MOList for retrieving omics data}
}
\value{
A list containing the ATAC peaks for condition 1 and condition 2
}
\description{
Getter for the ATAC peaks data from the MOList object
}
\examples{
\dontrun{
# Using the example MOList object
dataATACpeaks <- ATACpeaks(myMOList)
}
}
