% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOList.R
\name{ATACpeaks<-}
\alias{ATACpeaks<-}
\title{Setter for the ATACpeaks slot}
\usage{
ATACpeaks(x) <- value
}
\arguments{
\item{x}{An object of class MOList for appending/exchanging omics data}

\item{value}{A list containing the ATAC peaks for condition 1 and condition 2}
}
\value{
An object of class MOList
\itemize{
\item \code{RNAseq}: A numeric matrix containing the RNAseq data
\item \code{RNAseqSamples}: A list containing the sample names and grouping
                        information for the RNAseq data
\item \code{smallRNAseq}: A numeric matrix containing the smallRNAseq data
\item \code{smallRNAseqSamples}: A list containing the sample names and
                             grouping information for the smallRNAseq data
\item \code{proteomics}: A numeric matrix containing the proteomics data
\item \code{proteomicsSamples}: A list containing the sample names and
                          grouping information for the proteomics data
\item \code{ATACpeaks}: A list containing the ATAC peaks for condition 1 and
                  condition 2
}
}
\description{
Setter for the ATACpeaks slot
}
\examples{
\dontrun{
# Assuming peakCond1 and peakCond2 are data frames containing the ATAC peaks
# for condition 1 and condition 2, respectively

# Setting the ATAC peaks data to the MOList object
ATACpeaks(myMOList) <- list(peaksCond1 = peakCond1, peaksCond2 = peakCond2)
}

}
\keyword{internal}
