% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkSmallAnnoCoverage}
\alias{checkSmallAnnoCoverage}
\title{Check if the annotation covers all small RNA transcripts in the MOList
object}
\usage{
checkSmallAnnoCoverage(objMOList, anno)
}
\arguments{
\item{objMOList}{A object of class MOList}

\item{anno}{A character string indicating the annotation to be used. Either
"human" or "userAnno".}
}
\description{
This function checks if the annotation covers all small RNA
             transcripts in the MOList object. If not, it will throw an
             warning message, but the function will still continue to run.
}
\keyword{internal}
