% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructNetwork.R
\name{combineSncInteractions}
\alias{combineSncInteractions}
\title{Combine small RNA - mRNA interactions from different sources}
\usage{
combineSncInteractions(
  predInteract,
  extInteract,
  sncAnno,
  deResultRNA,
  deResultSmallRNA,
  smallRNATypes = SMALLRNA_CATEGORIES
)
}
\arguments{
\item{predInteract}{A list of predicted interactions in adjacent list format
\itemize{
\item \code{regulator}: A character vector containing the regulators
\item \code{target}: A character vector containing the targets
}}

\item{extInteract}{A list of curated interactions in adjacent list format}

\item{sncAnno}{A list of annotations for small RNAs}

\item{deResultRNA}{A data frame containing the differential RNAseq data}

\item{deResultSmallRNA}{A data frame containing the differential smallRNAseq}
}
\value{
A list of interactions in adjacent list format, with the interactions
        that are common to both lists, with the regulator type annotated,
        and with the miRNA - mRNA interactions filtered. If only one type
        of interactions is provided, then the function will return the
        interactions directly.
}
\description{
This function combines the small RNA - mRNA interactions from
             different sources, including the predicted interactions and the
             curated interactions. The function also annotates the type of
             the regulators. For miRNA only, the functions also filters the
             interactions by inverse correlation.
}
\keyword{internal}
