% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffOmicsAnalysis.R
\name{countDiffExpr}
\alias{countDiffExpr}
\title{Differential expression analysis of count-based omics data}
\usage{
countDiffExpr(objMOList, omic, batch, program = DESEQ2)
}
\arguments{
\item{objMOList}{A MOList object containing the omics data}

\item{omic}{A character string specifying the omics data to be analyzed
must be one of "RNAseq", "smallRNAseq", and "proteomics"}

\item{batch}{A character vector specifying the batch information for the
omics data, must be the same length as the number of samples in
the omics data, used for batch correction. Can be NULL if no
batch correction is needed.}

\item{program}{A character string specifying the program used for the
analysis, DESeq2 or EdgeR}
}
\value{
An MOList object containing the differential expression analysis
        results. Results are appended to the original MOList object as
        list elements named "DERNAseq", "DEsmallRNAseq", and "DEProteomics"
        for RNAseq, small RNAseq, and protein data, respectively.
\itemize{
\item \code{RNAseq}: A numeric matrix containing the RNAseq data
\item \code{RNAseqSamples}: A list containing the sample names and grouping
                        information for the RNAseq data
\item \code{smallRNAseq}: A numeric matrix containing the smallRNAseq data
\item \code{smallRNAseqSamples}: A list containing the sample names and
                             grouping information for the smallRNAseq data
\item \code{proteomics}: A numeric matrix containing the proteomics data
\item \code{proteomicsSamples}: A list containing the sample names and
                          grouping information for the proteomics data
\item \code{ATACpeaks}: A list containing the ATAC peaks for condition 1 and
                  condition 2
\item \code{DERNAseq}: A DETag object containing the differential
                       expression analysis results for the RNAseq data
\item \code{DEsmallRNAseq}: A DETag object containing the differential
                            expression analysis results for the smallRNAseq
                            data
\item \code{DEproteomics}: A DETag object containing the differential
                           expression analysis results for the proteomics
                           data
}
}
\description{
This function performs differential expression analysis on the
             count-based omics data. The RNAseq, small RNAseq, and protein
             data are supported for differential expression analysis.
             Calling this function again on the same omics data will
             overwrite the previous results.
}
\note{
Preconditions:
\itemize{
\item The input omics data must be a MOList object
\item The input omics data must be filtered by filterGeneCounts
\item The batch information must have the same length as the number of
      samples in the omics data
}
}
\examples{
# Example 1: Differential expression analysis of RNAseq data without batch
#            correction

# Create example count matrix
countMatrix <- matrix(sample(0:100, 1000, replace = TRUE),
  nrow = 100, ncol = 10
)

# Create example grouping information
group <- rep(c("A", "B"), each = 5)

# Create example MOList object
objMOList <- MOList(RNAseq = countMatrix, RNAGroupBy = group)

# Perform differential expression analysis
objMOList <- countDiffExpr(objMOList, "RNAseq", group)


# Example 2: Differential expression analysis of protein data with batch
#            correction

# Create example count matrix
proteinCounts <- matrix(sample(0:100, nrow = 100, ncol = 20, replace = TRUE))

# Create example grouping information
proteinGroup <- rep(c("A", "B"), each = 10)

# Create example batch information
proteinBatch <- c(rep("batch1", 5), rep("batch2", 15))

# Create example MOList object
objMOList <- MOList(
  RNAseq = countMatrix, RNAGroupBy = group,
  proteomics = proteinCounts,
  proteomicsGroupBy = proteinGroup
)

# Perform differential expression analysis
objMOList <- countDiffExpr(
  objMOList, "proteomics",
  proteinGroup, proteinBatch
)

}
\references{
Robinson MD, McCarthy DJ, Smyth GK. edgeR: a Bioconductor package for
differential expression analysis of digital gene expression data.
Bioinformatics. 2010 Jan 1;26(1):139-40. doi: 10.1093/bioinformatics/btp616.
Epub 2009 Nov 11. PMID: 19910308; PMCID: PMC2796818.

Love, M.I., Huber, W., and Anders, S. (2014). Moderated estimation of fold
change and dispersion for RNA-seq data with DESeq2. Genome Biology 15, 1–21.
}
