% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEOmics.R
\name{countPCA}
\alias{countPCA}
\title{Principle Component Analysis (PCA) for count-based omics data}
\usage{
countPCA(
  matrix,
  groupBy,
  batch = NULL,
  title = NULL,
  col1 = "#440154",
  col2 = "#fde725"
)
}
\arguments{
\item{matrix}{A numeric matrix containing the count-based omics data, must be
raw count expression matrix, not normalized by any method}

\item{groupBy}{A vector of factors for grouping samples, which should be
consistent with the order of samples in the count matrix.}

\item{batch}{A vector of factors for batch effect, which should be consistent
with the order of samples in the count matrix. Default is NULL.}

\item{title}{A character string specifying the title for the plot, default
is NULL.}

\item{col1}{A string specifying the color for the first group, which should
be a valid color name recognized by R or a hexadecimal color
code. Default is "#440154".}

\item{col2}{The color for the second group, default is "#fde725". If
continuous grouping variable is used, the color gradient will be
used instead of the two colors}
}
\value{
A ggplot object
}
\description{
This function performs PCA analysis for count-based omics data.
             The user can specify the title for the plot.
}
\details{
Raw counts are normalized by RLE method and then transformed by
         variance stabilizing transformation (VST) before PCA analysis.
}
\examples{
# Use the package-provided example data
data("expMOList")
data(RNAseq_heart_samples)

# Extract the information needed for PCA analysis for example RNAseq data
matrixRNAseq <- getRawData(expMOList, "RNAseq")
groupBy <- RNAseq_heart_samples$Age

# Example 1: Perform PCA analysis without batch effect correction
pcaPlot <- countPCA(matrixRNAseq, groupBy)

pcaPlot

# Example 2: Perform PCA analysis with batch effect correction
# Create an example batch effect
batch <- rep(c("Batch1", "Batch2"), each = 30)[1:53]
pcaPlot <- countPCA(matrixRNAseq, groupBy, batch = batch)

pcaPlot

}
\references{
\insertRef{villanueva2019ggplot2}{IntegraTRN}

\insertRef{dplyr}{IntegraTRN}

\insertRef{love2014moderated}{IntegraTRN}
}
