% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{csAnnoToDF}
\alias{csAnnoToDF}
\title{Format csAnno object to a data frame}
\usage{
csAnnoToDF(anno)
}
\arguments{
\item{anno}{A csAnno object}
}
\value{
A data frame containing the annotation information
}
\description{
Format csAnno object to a data frame
}
\keyword{internal}
