% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DESeqDesign}
\alias{DESeqDesign}
\title{Defining DESeq2 colData and design from vectors}
\usage{
DESeqDesign(groupBy, batch = NULL)
}
\arguments{
\item{groupBy}{A vector of group names}

\item{batch}{A vector of batch names}
}
\value{
A list containing the colData and design, where colData is a data
        frame and design is a formula
}
\description{
Defining DESeq2 colData and design from vectors
}
\keyword{internal}
