% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DETag.R
\name{DETag}
\alias{DETag}
\title{Constructor for the DETag class}
\usage{
DETag(DEResult, method, normalizedCounts = NULL)
}
\arguments{
\item{DEResult}{A data frame containing the differential expression analysis
results}

\item{method}{A character string specifying the program used for the analysis}

\item{normalizedCounts}{A matrix containing the normalized counts for the
genes, with each row representing a gene and each
column representing a sample}
}
\value{
An object of the DETag class
\itemize{
\item \code{DEResult}: A data frame containing the differential expression
                       analysis results
\item \code{method}: A character string specifying the program used for the
                     analysis
}
}
\description{
This function creates an object of the DETag class
}
\examples{
# Create an example data frame
deResult <- data.frame(
  gene = paste0("gene", seq_len(10)),
  logFC = runif(10),
  adj.P.Val = runif(10)
)

# Create a normalized counts matrix
normalizedCounts <- matrix(runif(100), nrow = 10)

# Create an object of the DETag class
deTag <- DETag(deResult, "DESeq2", normalizedCounts)

# Check the class of the object
class(deTag)

}
