% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffOmicsAnalysis.R
\docType{class}
\name{DETags-class}
\alias{DETags-class}
\title{DETags S4 class}
\description{
This class is used to store the differential expression analysis
             results for the count-based omics data. Although the
             functionality of the class itself can be easily implemented
             by using other data structures, here use specifically designed
             S4 class to allow future extension of the program, e.g. 
             providing different methods for differential expression
             analysis while maintaining a consistent interface.
}
\section{Slots}{

\describe{
\item{\code{DEResult}}{A data frame containing the differential expression analysis
results}

\item{\code{method}}{A character string specifying the program used for the analysis}
}}

\references{
Robinson MD, McCarthy DJ, Smyth GK. edgeR: a Bioconductor package for
differential expression analysis of digital gene expression data.

Love, M.I., Huber, W., and Anders, S. (2014). Moderated estimation of fold 
change and dispersion for RNA-seq data with DESeq2. Genome Biology 15, 1–21.

Advanced R by H. Wickham. Access: https://adv-r.hadley.nz/index.html
}
\keyword{internal}
