% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffOmicsAnalysis.R
\name{DETags}
\alias{DETags}
\title{Constructor for the DETags class}
\usage{
DETags(DEResult, method)
}
\arguments{
\item{DEResult}{A data frame containing the differential expression analysis
results}

\item{method}{A character string specifying the program used for the analysis}
}
\value{
An object of the DETags class
\itemize{
\item \code{DEResult}: A data frame containing the differential expression
                       analysis results
\item \code{method}: A character string specifying the program used for the
                     analysis
}
}
\description{
This function creates an object of the DETags class
}
\examples{
# Create an example data frame
deResult <- data.frame(
 gene = paste0("gene", seq_len(10)),
logFC = runif(10),
adj.P.Val = runif(10)
)

# Create an object of the DETags class
detags <- DETags(deResult, "DESeq2")

# Check the class of the object
class(detags)

}
\references{
Advanced R by H. Wickham. Access: https://adv-r.hadley.nz/index.html
}
\keyword{internal}
