% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffOmicsAnalysis.R
\name{diffOmicsAnalysis}
\alias{diffOmicsAnalysis}
\title{Perform differential analysis on the omics data}
\usage{
diffOmicsAnalysis(
  objMOList,
  rnaseqBatch = NULL,
  smallRnaBatch = NULL,
  proteinBatch = NULL
)
}
\arguments{
\item{objMOList}{A MOList object containing the omics data}

\item{rnaseqBatch}{A character vector specifying the batch information for
the RNA-seq data, must be the same length as the number
of samples in the RNA-seq data, used for batch correction}

\item{smallRnaBatch}{A character vector specifying the batch information for
the small RNA data, must be the same length as the
number of samples in the small RNAseq data, used for
batch correction}

\item{proteinBatch}{A character vector specifying the batch information for
the protein data, must be the same length as the number
of samples in the protein data, used for batch correction}
}
\value{
An MOList object containing the differential analysis results
}
\description{
This function performs data filtering, normalization, batch
             correction, and differential analysis on each of the omics data.
             RNAseq, small RNAseq, and protein data are supported for
             differential expression analysis. The ATACseq data are used
             used to calculate the differential accessible regions.
             While performing the differential analysis, the original count
             data for the RNAseq, small RNAseq, and protein data are
             filtered and subsequently normalized by library size. After
             calling diffOmicsAnalysis, the users can use the S4 methods
             of the MOList class to extract filtered and normalized read
             counts for each of the omics data.
}
