% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processPeaks.R
\name{enrichMotifs}
\alias{enrichMotifs}
\title{Perform motif enrichment analysis on peaks with binary conditions}
\usage{
enrichMotifs(objMOList, bsgenome, pwmL, fixedWidth = 500)
}
\arguments{
\item{objMOList}{An object of class MOList}

\item{bsgenome}{The BSgenome object that contains the genome sequence
Read the Bioconductor
vignette for the BSgenome package for more details.}

\item{pwmL}{A PWMatrixList object containing the position-weight matrices
(PWMs) to use for motif enrichment analysis. The PWMs can be
generated using the getMatrixSet() function from the TFBSTools
package. See the vignette of the TFBSTools package for more
details.}

\item{fixedWidth}{The fixed width to adjust the peak size for motif if
necessary (default: 500). The default width follows an
optimal width peak selection strategy defined by
\insertCite{grandi2022chromatin;textual}{IntegraTRN}.}
}
\description{
This function performs motif enrichment analysis on peaks with
             binary conditions. The motif enrichment analysis is performed
             using the monaLisa package.
}
\examples{
\dontrun{
# Use the package-provided example data
data("expMOList")

# Retrieve genome data
bsgenome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38
# Generate position-weight matrices (PWMs) from the JASPAR database
pwmL <- TFBSTools::getMatrixSet(JASPAR2022::JASPAR2022,
  opts = list(
    matrixtype = "PWM",
    tax_group = "vertebrates"
  )
)
# Perform motif enrichment analysis
expMOList <- enrichMotifs(expMOList,
  bsgenome = bsgenome,
  pwmL = pwmL,
  fixedWidth = 500
)
}
}
\references{
\insertRef{BSgenome}{IntegraTRN}

\insertRef{tan2016tfbstools}{IntegraTRN}

\insertRef{grandi2022chromatin}{IntegraTRN}

\insertRef{machlab2022monalisa}{IntegraTRN}

\insertRef{Biostrings}{IntegraTRN}
}
\keyword{internal}
