% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOPTag.R
\name{exportDE,TOPTag-method}
\alias{exportDE,TOPTag-method}
\title{Export the differential analysis results from a TopTag object}
\usage{
\S4method{exportDE}{TOPTag}(x, original = FALSE)
}
\arguments{
\item{x}{An object of the DETag class}

\item{original}{A boolean indicating whether to return the original results.
If FALSE, the results will be converted to a data frame with
the following columns: logFC, pvalue, padj
(default: FALSE)}
}
\value{
A data frame containing the differential analysis results.
}
\description{
This function exports the differential analysis results from a
             TopTag object. Overwrites the exportDE function in the DETag
             class due to different nomenclature of the columns.
}
\examples{
# Use the package provided example data
data("expMOList")

# Extract an DETag object from RNA-seq data
deTag <- expMOList$DERNAseq

# Create a TOPTag object
topTag <- TOPTag(deTag, logFCCutoff = 0, topGenes = 0.2)

# Export the differential analysis results
exportDE(topTag)

}
