% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOList.R
\name{exportDiffGenes}
\alias{exportDiffGenes}
\title{Export a list of differentially expressed genes}
\usage{
exportDiffGenes(objMOList, experiment, outPath, log2FC = 1, adjP = 0.05)
}
\arguments{
\item{objMOList}{A MOList object containing the omics data}

\item{experiment}{A character string specifying the experiment type, must
be one of "RNAseq", "smallRNAseq", and "proteomics"}

\item{outPath}{A character string specifying the output file path}

\item{log2FC}{A numeric value specifying the cutoff for the log2 fold change
of the differentially expressed genes, default is 1}

\item{adjP}{A numeric value specifying the cutoff for the adjusted p-value
of the differentially expressed genes, default is 0.05}
}
\description{
This function exports a list of differentially expressed genes
             from the MOList object. The user can specify the experiment type
             and the cutoffs for the log2 fold change and adjusted p-value.
             The function will export a file containing the differentially
             expressed genes as a comma- or tab-separated text file. The
             uses the suffix ".csv" or ".txt/tsv" to determine the file type.
}
\examples{
# Assuming myMOList contains the differential omics data
# i.e., diffOmics has been performed

# Example 1: Exporting the differentially expressed genes from RNAseq
\dontrun{
exportDiffGenes(myMOList, "RNAseq", "diffGenes_RNAseq.txt")
}

# Example 2: Exporting the differentially expressed genes from smallRNAseq
# with custom cutoffs
\dontrun{
exportDiffGenes(myMOList, "smallRNAseq", "diffGenes_smallRNAseq.txt",
  log2FC = 2, adjP = 0.01)
}

}
