% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRNet.R
\name{exportEdgeSet}
\alias{exportEdgeSet}
\alias{exportEdgeSet,TRNet-method}
\title{Export all network interactions from TRNet object}
\usage{
exportEdgeSet(trn)
}
\arguments{
\item{trn}{A TRNet object}
}
\value{
A data frame containing all network interactions
}
\description{
This function exports all network interactions from TRNet object
}
\examples{
# Define some example edges
edges <- data.frame(
  regulator = c("A", "B", "C"),
  target = c("D", "E", "F"),
  regulatorType = c("miRNA", "TF", "TF")
)

# Create TRNet object
trn <- TRNet(edges, FALSE, "RNA-seq")

# Export all network interactions
exportEdgeSet(trn)

}
