% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRNet.R
\name{exportIgraph}
\alias{exportIgraph}
\title{Export an igraph object from TRNet object}
\usage{
exportIgraph(trn)
}
\arguments{
\item{trn}{A TRNet object}
}
\value{
An igraph object
}
\description{
This function exports an igraph object from TRNet object
}
\examples{
# Define some example edges
edges <- data.frame(
  regulator = c("A", "B", "C"),
  target = c("D", "E", "F"),
  regulatorType = c("miRNA", "TF", "TF")
)

# Create TRNet object
trn <- TRNet(edges, FALSE, "RNA-seq")

# Export igraph object
exportIgraph(trn)

}
