% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRNAsmallRNAcoExprEstimate.R
\name{exportMatchResult}
\alias{exportMatchResult}
\title{Export the match result of RNAseq and small RNAseq samples}
\usage{
exportMatchResult(objMOList)
}
\arguments{
\item{objMOList}{An object of class MOList}
}
\value{
A data frame containing the match result, with the following format:
\itemize{
\item{RNAseq}{The sample names of the RNAseq samples.}
\item{SmallRNAseq}{The sample names of the small RNAseq samples.}
}
}
\description{
This function exports the match result of RNAseq and small
             RNAseq samples.
}
\examples{
# Suppose we have an MOList object with RNAseq and small RNAseq data
\dontrun{
# Perform matching
myMOList <- matchSamplesRNAsmallRNA(myMOList)
# Export the match result
exportMatchResult(myMOList)
}

}
