% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processSmallRNA.R
\name{extractTranscriptFromAnno}
\alias{extractTranscriptFromAnno}
\title{Extract small RNA categories}
\usage{
extractTranscriptFromAnno(annoDF, category)
}
\arguments{
\item{annoDF}{A data frame containing the annotation information for small
RNA transcripts. The first column should be the transcript name
of the small RNAs, and the second column should be the category
of the small RNAs.}

\item{category}{A character string indicating the category of small RNAs to
be extracted.}
}
\value{
A character vector containing the small RNA transcripts belonging to
       the specified category.
}
\description{
This function extracts the small RNAs belonging to each of the
             categories, based on input.
}
\references{
\insertRef{dplyr}{IntegraTRN}
}
\keyword{internal}
