% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructNetwork.R
\name{filtermiRNAinverseCorr}
\alias{filtermiRNAinverseCorr}
\title{Select miRNA inverse correlation}
\usage{
filtermiRNAinverseCorr(
  exprAdjList,
  deResultRNA,
  deResultSmallRNA,
  smallRNATypes = SMALLRNA_CATEGORIES
)
}
\arguments{
\item{exprAdjList}{A list representing the regulator-target interactions
\itemize{
\item \code{regulator}: A character vector containing the regulators
\item \code{target}: A character vector containing the targets
}}

\item{smallRNATypes}{A character vector containing the small RNA types that
the user wants to use to construct the network. The
available types are "miRNA", "piRNA", "snRNA", "snoRNA",
"tRNA", and "circRNA".}

\item{rnaDETag}{A DETag object containing the differential RNAseq data}

\item{smallDETag}{A DETag object containing the differential smallRNAseq data}
}
\value{
A list with filtered regulator-target interactions
}
\description{
This function selects the miRNAs in which its expression is
             inversely correlated with the expression of the target gene.
}
