% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructNetwork.R
\name{filterTargetGenes}
\alias{filterTargetGenes}
\title{Filter target genes by proteomics data}
\usage{
filterTargetGenes(rnaTopTag, protTopTag, mapping)
}
\arguments{
\item{rnaTopTag}{A TOPTag object containing the top differential genes from
the RNAseq data}

\item{protTopTag}{A TOPTag object containing the top differential genes from
the proteomics data}

\item{mapping}{A data frame containing the mapping between genes and proteins
\itemize{
\item \code{gene}: A character vector containing the gene names
\item \code{protein}: A character vector containing the protein names
}}
}
\value{
A TOPTag object containing the top differential genes from the
        RNAseq data, with the target genes filtered by proteomics data
}
\description{
This function filters the target genes by proteomics data.
             Only genes that shows consistent expression changes in both
             RNAseq and proteomics data will be used to construct the
             network. This is to ensure high confidence in the target genes.
}
\keyword{internal}
