% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findGeneType}
\alias{findGeneType}
\title{Annotate the type of gene}
\usage{
findGeneType(geneName, annotation)
}
\arguments{
\item{geneName}{The name of a gene}

\item{annotation}{A list, with the name of each element defined to be the
type of gene, and the content of each element defined to be
a vector of gene names in that type}
}
\value{
The type of the gene
}
\description{
Annotate the type of gene
}
\keyword{internal}
