% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getGroupingInfo}
\alias{getGroupingInfo}
\title{Retrieving the grouping information for a specific omics data}
\usage{
getGroupingInfo(objMOList, experiment)
}
\arguments{
\item{objMOList}{A MOList object containing the omics data}

\item{experiment}{A character string specifying the experiment type, must
be one of "RNAseq", "smallRNAseq", and "proteomics"}
}
\value{
A vector of grouping information
}
\description{
Retrieving the grouping information for a specific omics data
}
\examples{
\dontrun{
# Assuming myMOList is a MOList object

# Get the grouping information for the RNAseq data
getGroupingInfo(myMOList, "RNAseq")

# Get the grouping information for the smallRNAseq data
getGroupingInfo(myMOList, "smallRNAseq")

# Get the grouping information for the proteomics data
getGroupingInfo(myMOList, "proteomics")
}

}
\keyword{internal}
