% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOList.R
\name{getRawData,MOList-method}
\alias{getRawData,MOList-method}
\alias{getRawData}
\title{Getter for the count-based omics data from the MOList object}
\usage{
\S4method{getRawData}{MOList}(x, omics)
}
\arguments{
\item{x}{An object of class MOList for retrieving omics data}

\item{omics}{A character string specifying the omics data to be retrieved
must be one of "RNAseq", "smallRNAseq", "proteomics", or "ATAC"}
}
\value{
Raw data of the specified omics type: a numeric matrix for
        count-based omics data, and a list of data frames for ATAC peaks
}
\description{
This function is a getter for the count-based omics data from
             the MOList object. The user can retrieve the RNAseq,
             smallRNAseq, and proteomics data from the MOList object.
}
\examples{
\dontrun{
# Using the example MOList object
dataRNAseq <- getCounts(myMOList, "RNAseq")
dataSmallRNAseq <- getCounts(myMOList, "smallRNAseq")
dataProteomics <- getCounts(myMOList, "proteomics")
dataATAC <- getCounts(myMOList, "ATAC")
}

}
