% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOList.R
\name{getRawData}
\alias{getRawData}
\alias{getRawData,MOList-method}
\title{Getter for the count-based omics data from the MOList object}
\usage{
getRawData(x, omics)
}
\arguments{
\item{x}{An object of class MOList for retrieving omics data}

\item{omics}{A character string specifying the omics data to be retrieved
must be one of "RNAseq", "smallRNAseq", "proteomics", or "ATAC"}
}
\value{
Raw data of the specified omics type: a numeric matrix for
        count-based omics data, and a list of data frames for ATAC peaks
        that are differentially accessible between two conditions
}
\description{
This function is a getter for the count-based omics data from
             the MOList object. The user can retrieve the RNAseq,
             smallRNAseq, and proteomics data from the MOList object.
}
\examples{
# Use the package-provided example MOList object
data("expMOList")

# Example 1: Retrieving the RNAseq raw counts
getRawData(expMOList, omics = "RNAseq")

# Example 2: Retrieving the smallRNAseq raw counts
getRawData(expMOList, omics = "smallRNAseq")

# Example 3: Retrieving the proteomics raw counts
getRawData(expMOList, omics = "proteomics")

# Example 4: Retrieving the ATAC peaks as a list
getRawData(expMOList, omics = "ATAC")

}
