% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOList.R
\name{getSampleInfo,MOList-method}
\alias{getSampleInfo,MOList-method}
\title{Retrieving the grouping information for a specific omics data}
\usage{
\S4method{getSampleInfo}{MOList}(x, experiment)
}
\arguments{
\item{x}{A MOList object containing the omics data}

\item{experiment}{A character string specifying the experiment type, must
be one of "RNAseq", "smallRNAseq", and "proteomics"}
}
\value{
A vector of grouping information
}
\description{
Retrieving the grouping information for a specific omics data
}
\examples{
\dontrun{
# Assuming myMOList is a MOList object

# Get the grouping information for the RNAseq data
getSampleInfo(myMOList, "RNAseq")

# Get the grouping information for the smallRNAseq data
getSampleInfo(myMOList, "smallRNAseq")

# Get the grouping information for the proteomics data
getSampleInfo(myMOList, "proteomics")
}

}
\keyword{internal}
