% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSampleNames}
\alias{getSampleNames}
\title{Identify the sample names from a count matrix}
\usage{
getSampleNames(countMatrix)
}
\arguments{
\item{countMatrix}{A count matrix with samples as columns and features as
rows}
}
\value{
A vector of sample names
}
\description{
Identify the sample names from a count matrix
}
\keyword{internal}
