% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegraTRN-package.R
\docType{package}
\name{IntegraTRN-package}
\alias{IntegraTRN}
\alias{IntegraTRN-package}
\title{IntegraTRN: Integrated Omics For Inference Of Transcriptional Regulatory Network}
\description{
The R package 'IntegraTRN' integrates transcriptomic, small RNAomic, proteomic, and epigenomic data to construct a transcriptional regulatory network (TRN) underlying the gene expression changes in a developmental or disease (or any continuous/binary) biological context. In particular, the TRN is a network of interacting factors, in which one element can exert a regulatory effect (activating or inhibitory) on the expression of one or more elements. Given the complex nature of transcriptional regulation, the package reveals the key players of such regulation, including transcriptional factors (TFs) and small RNAs. Since gene/protein expression changes are usually the most direct molecular cause of observed phenotypic alterations, the core TRN deciphers changes between two or more conditions and infers the upstream regulatory factors that mediate such changes. The analysis pipeline provided by this package is primarily composed of a two-step process: (1) elucidating the transcriptional and regulatory changes that take place during a biological process, such as development or disease; and (2) performing correlational analysis with rigorous filtering based on biological data to identify the regulatory interactions that are responsible for the observed changes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/j-y26/IntegraTRN}
  \item Report bugs at \url{https://github.com/j-y26/IntegraTRN/issues}
}

}
\author{
\strong{Maintainer}: Jielin Yang \email{jielin.yang@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-4207-828X}{ORCID})

}
\keyword{internal}
