% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructNetwork.R
\name{intersectInteractions}
\alias{intersectInteractions}
\title{Intersect interaction lists}
\usage{
intersectInteractions(adjList1, adjList2)
}
\arguments{
\item{adjList1}{A list of interactions in adjacent list format
\itemize{
\item \code{regulator}: A character vector containing the regulators
\item \code{target}: A character vector containing the targets
}}

\item{adjList2}{A list of interactions in adjacent list format}
}
\value{
A list of interactions in adjacent list format, with the interactions
        that are common to both lists
}
\description{
This function intersects two lists of interactions in
             adjacent list format.
}
\note{
At least one of the lists must be non-empty.
}
