% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{labelSmallSizeGroup}
\alias{labelSmallSizeGroup}
\title{Label the samples with smaller sample size as 1}
\usage{
labelSmallSizeGroup(sampleDF, identifier, colname)
}
\arguments{
\item{sampleDF}{A data frame containing the sample names as row names}

\item{identifier}{A string to identify the samples in sample names}

\item{colname}{The column name to be added to the data frame}
}
\value{
A data frame with the column added
}
\description{
Label the samples with smaller sample size as 1
}
\keyword{internal}
