% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRNAsmallRNAcoExprEstimate.R
\name{matchBinary}
\alias{matchBinary}
\title{Match RNAseq and small RNAseq samples based on binary grouping variable}
\usage{
matchBinary(sampleDF)
}
\arguments{
\item{sampleDF}{A data frame containing the sample information.
                Must follow the following format:
\itemize{
\item{row.names}{The sample names that correspond to the sample names in
                 the MOList object, with a suffix "_smallRNA" for small
                 RNAseq samples.}
\item{groupBy}{The grouping variable for matching, must be binary.}
}}
}
\value{
A data frame containing the matched sample information, with the
       following format:
\itemize{
\item{smallRNA}{The sample names of the small RNAseq samples. No suffix.}
\item{RNA}{The sample names of the RNAseq samples.}
}
}
\description{
This function performs optimal pair matching, a variant of
             nearest neighbor marching for RNAseq and small RNAseq samples
             based on a binary grouping variable. The matching is performed
             separately for the two groups.
}
\references{
\insertRef{stuart2011matchit}{IntegraTRN}

\insertRef{gong2020optmatch}{IntegraTRN}
}
\keyword{internal}
