% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRNAsmallRNAcoExprEstimate.R
\name{matchContinuous}
\alias{matchContinuous}
\title{Match RNAseq and small RNAseq samples based on continuous grouping variable}
\usage{
matchContinuous(sampleDF)
}
\arguments{
\item{sampleDF}{A data frame containing the sample information. Must follow
                the following format:
\itemize{
 \item{row.names}{The sample names that correspond to the sample names in
                  the MOList object, with a suffix "_smallRNA" for small
                  RNAseq samples.}
 \item{groupBy}{The grouping variable for matching.}
}}
}
\value{
A data frame containing the matched sample information, with the
        following format:
\itemize{
\item{smallRNA}{The sample names of the small RNAseq samples. No suffix.}
\item{RNA}{The sample names of the RNAseq samples.}
}
}
\description{
This function performs optimal pair matching, a variant of
             nearest neighbor marching for RNAseq and small RNAseq samples
             based on a continuous grouping variable.
}
\references{
\insertRef{stuart2011matchit}{IntegraTRN}

\insertRef{gong2020optmatch}{IntegraTRN}
}
\keyword{internal}
