% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRNAsmallRNAcoExprEstimate.R
\name{matchInGroup}
\alias{matchInGroup}
\title{Match RNAseq and small RNAseq samples within a single group}
\usage{
matchInGroup(sampleDF)
}
\arguments{
\item{sampleDF}{A data frame containing the sample information. Must follow
               the following format:
\itemize{
\item{row.names}{The sample names that correspond to the sample names in
                 the MOList object, with a suffix "_smallRNA" for small
                 RNAseq samples.}
\item{groupBy}{The grouping variable for matching.}
}}
}
\value{
A data frame containing the matched sample information, with the
        following format:
\itemize{
\item{smallRNA}{The sample names of the small RNAseq samples. No suffix.}
\item{RNA}{The sample names of the RNAseq samples.}
}
}
\description{
This function performs optimal pair matching, a variant of
             nearest neighbor marching for RNAseq and small RNAseq samples
             within a single group defined from binary classification
}
\references{
\insertRef{stuart2011matchit}{IntegraTRN}
}
\keyword{internal}
