% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matchResultToDF}
\alias{matchResultToDF}
\title{Transform the result of MatchIt::matchit() to a data frame}
\usage{
matchResultToDF(matchResult)
}
\arguments{
\item{matchResult}{The result of MatchIt::matchit()}
}
\value{
A data frame containing the matched pairs
}
\description{
Transform the result of MatchIt::matchit() to a data frame
}
\keyword{internal}
