% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRNAsmallRNAcoExprEstimate.R
\name{matchSamplesRNAsmallRNA}
\alias{matchSamplesRNAsmallRNA}
\title{Match RNAseq and small RNAseq samples}
\usage{
matchSamplesRNAsmallRNA(
  objMOList,
  sampleDFRNAseq = NULL,
  sampleDFSmallRNAseq = NULL,
  varMatch = NULL
)
}
\arguments{
\item{objMOList}{An object of class MOList}

\item{sampleDFRNAseq}{A data frame containing the RNAseq sample information.
Each sample is a row in the data frame, with row names
being the sample names that correspond to the sample
names in the MOList object.}

\item{sampleDFSmallRNAseq}{A data frame containing the small RNAseq sample
information. Each sample is a row in the data
frame, with row names being the sample names that
correspond to the sample names in the MOList
object.}

\item{varMatch}{A vector of variable names that will be used for matching.
The variable names should be present in both sampleDFRNAseq
and sampleDFSmallRNAseq. If NULL, default to use all
common variables in the sample data frames.}
}
\value{
An object of class MOList with the matching information stored in
        a list of two numeric vectors with equal length, where each element
        in the first vector is the index of the RNAseq sample, and the
        corresponding element in the second vector is the index of the
        matched small RNAseq sample. These indices match to the sample names
        in the RNAseqSamples and SmallRNAseqSamples slots of the MOList
        object.
}
\description{
This function generates a one-to-one matching between the
             RNAseq samples and the small RNAseq samples using the
             optimal pair matching algorithm, which is a variant of nearest
             neighbor matching. The matching is based on the grouping
             variable and other variables in the sample data frames.
}
