% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matchVecToMatrix}
\alias{matchVecToMatrix}
\title{Match vector length to matrix column length}
\usage{
matchVecToMatrix(vec, mat)
}
\arguments{
\item{vector}{A vector}

\item{matrix}{A matrix}
}
\value{
A boolean indicating whether the vector length matches the matrix
        column length
}
\description{
Match vector length to matrix column length
}
\examples{
# Create example vector and matrix
vec <- 1:10
mat <- matrix(1:100, nrow = 10, ncol = 10)

# Match vector length to matrix column length
matchVecToMatrix(vec, mat)

}
\keyword{internal}
