% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOList.R
\name{MOList}
\alias{MOList}
\title{Constructor for the MOList object}
\usage{
MOList(
  objMOList = NA,
  RNAseq = NA_real_,
  RNAGroupBy = NA,
  smallRNAseq = NA_real_,
  smallRNAGroupBy = NA,
  proteomics = NA_real_,
  proteomicsGroupBy = NA,
  pathATACpeak1 = NA,
  pathATACpeak2 = NA
)
}
\arguments{
\item{objMOList}{An object of class MOList for appending/exchanging omics
data}

\item{RNAseq}{A numeric matrix containing the RNAseq data, with row names as
gene names and column names as sample names}

\item{RNAGroupBy}{A vector of grouping information for the RNAseq data, used
to perform differential expression analysis}

\item{smallRNAseq}{A numeric matrix containing the smallRNAseq data, with row
names as gene names and column names as sample names}

\item{smallRNAGroupBy}{A vector of grouping information for the smallRNAseq
data, used to perform differential expression analysis}

\item{proteomics}{A numeric matrix containing the proteomics data, with row
names as gene names and column names as sample names}

\item{proteomicsGroupBy}{A vector of grouping information for the proteomics
data, used to perform differential expression
analysis}

\item{pathATACpeak1}{The path to a BED file containing the ATAC peaks for
condition 1}

\item{pathATACpeak2}{The path to a BED file containing the ATAC peaks for
condition 2}
}
\value{
An object of class MOList
}
\description{
This function is a constructor of the MOList object, which is
             used to store the multi-omics data. The MOList object can be
             constructed by providing the RNAseq, smallRNAseq,
             proteomics, and ATAC peaks data at once, or appending/exchanging
             each omics data separately. The RNAseq data is required for the
             initial construction of the MOList object. If data of existing
             omics type is provided, old data will be replaced.
}
