% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{motifNames}
\alias{motifNames}
\title{Get the names of enriched TFs}
\usage{
motifNames(enrichedMotifs)
}
\arguments{
\item{enrichedMotifs}{A SummarizedExperiment object containing the motif
enrichment results}
}
\value{
A vector of enriched TF names
}
\description{
Get the names of enriched TFs
}
\keyword{internal}
