% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{myMOList}
\alias{myMOList}
\title{Example MOList S4 object}
\format{
An S4 object of class MOList
\describe{
\item{RNAseq}{A matrix of RNAseq gene expression values}
\item{smallRNAseq}{A matrix of small RNAseq gene expression values}
\item{proteomics}{A matrix of proteomics gene expression values}
\item{ATACseq}{A list of two data frames of ATACseq peaks for 2 conditions}
}
}
\source{
Paul Delgado Olguin Lab, The Hospital for Sick Children, Toronto,
ON, Canada
}
\usage{
myMOList
}
\description{
An example MOList S4 object containing the RNAseq, small RNAseq, proteomics,
and ATACseq peaks
}
\examples{
\dontrun{
myMOList
}
}
\keyword{datasets}
