% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRNAsmallRNAcoExprEstimate.R
\name{nnRNAMatch}
\alias{nnRNAMatch}
\title{Performing optimal pair matching for RNAseq and small RNAseq samples}
\usage{
nnRNAMatch(sampleDFRNAseq, sampleDFSmallRNAseq)
}
\arguments{
\item{sampleDFRNAseq}{A data frame containing the RNAseq sample information.
Each sample is a row in the data frame, with row names
being the sample names that correspond to the sample
names in the MOList object.}

\item{sampleDFSmallRNAseq}{A data frame containing the small RNAseq sample
information. Each sample is a row in the data
frame, with row names being the sample names that
correspond to the sample names in the MOList
object.}
}
\value{
A list of two numeric vectors with equal length, where each element
        in the first vector is the index of the RNAseq sample, and the
        corresponding element in the second vector is the index of the
        matched small RNAseq sample.
}
\description{
This function performs optimal pair matching, a variant of
             nearest neighbour marching for RNAseq and
             small RNAseq samples using two sample data frames. The two data
             frames should contain the same number of variables, including
             the main grouping variable for DE analysis.
}
\details{
Depending on the nature of the grouping variables, either continuous
         or binary, the matching algorithm will be different. For continuous
         grouping variables, the grouping variable will has a weight equal to
         the number of other variables in the matching process. This is to
         ensure that the algorithm matches samples with the same grouping
         variable first, and then matches samples with similar values of the
         additional parameters. For binary grouping variables, the algorithm
         forces the matching process to be within the same group, and then
         performs optimal pair matching on other parameters.
}
