% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRNet.R
\name{parseVertexMetadata}
\alias{parseVertexMetadata}
\alias{parseVertexMetadata,TRNet-method}
\title{Parse vertex metadata from TRNet network metadata}
\usage{
parseVertexMetadata(trn)
}
\arguments{
\item{trn}{A TRNet object}
}
\value{
A data frame containing the vertex metadata
}
\description{
This function parses vertex metadata from the TRNet network
             metadata and returns the type of each vertex
}
\examples{
# Define some example edges
edges <- data.frame(
  regulator = c("A", "B", "C"),
  target = c("D", "E", "F"),
  regulatorType = c("miRNA", "TF", "TF")
)

# Create TRNet object
trn <- TRNet(edges, FALSE, "RNA-seq")

# Parse vertex metadata
parseVertexMetadata(trn)

}
