% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEAKTag.R
\name{PEAKTag}
\alias{PEAKTag}
\title{Constructor for the PEAKTag class}
\usage{
PEAKTag(objDETag, annotatedPeaks = NULL, TxDB = NULL, annoDB = NULL)
}
\arguments{
\item{objDETag}{A DETag object}

\item{annotatedPeaks}{A csAnno object containing the annotated peaks, as
defined in the ChIPseeker package}

\item{TxDB}{A TxDb object containing the transcript database used for the
annotation}

\item{annoDB}{A string indicating the annotation database used for the
annotation}
}
\value{
A PEAKTag object
}
\description{
This function constructs a PEAKTag object from a DETag object
             and the annotated peaks, as defined in the ChIPseeker package.
}
\examples{
# Make use of the package's provided example data
data(expMOList)

# Retrieve the DETag object for the ATACseq peaks
atacDETag <- expMOList$DEATAC

# Example 1: Construct a PEAKTag object without annotation
atacPEAKTag <- PEAKTag(atacDETag)

# Example 2: Construct a PEAKTag object with annotation
# Suppose TxDB and annoDB are already loaded
\dontrun{
atacPEAKTag <- PEAKTag(atacDETag, TxDB = TxDB, annoDB = "org.Hs.eg.db")
}

}
\references{
\insertRef{yu2015chipseeker}{IntegraTRN}

\insertRef{TxDb.Hsapiens.UCSC.hg38.knownGene}{IntegraTRN}

\insertRef{org.Hs.eg.db}{IntegraTRN}
}
