% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEOmics.R
\name{plotATACAnno}
\alias{plotATACAnno}
\title{Plotting the annotation Pie chart of differential accessible regions}
\usage{
plotATACAnno(objMOList)
}
\arguments{
\item{objMOList}{An MOList object containing the differential accessible
regions}
}
\value{
A ggplot object
}
\description{
This function generates a pie chart for the annotation of
             differential accessible regions based on the ATACseq data.
             This is a wrapper function for the plotAnnoPie function in the
             ChIPseeker package.
}
\examples{
\dontrun{
# Use the package-provided example data
data("expMOList")

# Annotates the peaks in the example data and performs motif enrichment
# Using human annotations as an example
library("TxDb.Hsapiens.UCSC.hg38.knownGene")
library("org.Hs.eg.db")
library("BSgenome.Hsapiens.UCSC.hg38")
txdb <- TxDb.Hsapiens.UCSC.hg38.knownGene
annoDb <- "org.Hs.eg.db"
bsgenome <- BSgenome.Hsapiens.UCSC.hg38

# Load PWMs from JASPAR
data("jasparVertebratePWM")

# Performs the annotation and motif enrichment
expMOList <- annotateATACPeaksMotif(expMOList,
  tssRegion = c(-3000, 3000),
  TxDb = txdb,
  annoDb = annoDb,
  bsgenome = bsgenome,
  pwmL = jasparVertebratePWM
)

# Plotting the annotation pie chart
plotATACAnno(expMOList)
}

}
\references{
\insertRef{yu2015chipseeker}{IntegraTRN}
}
