% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEOmics.R
\name{plotATACMotifHeatmap}
\alias{plotATACMotifHeatmap}
\title{Motif heatmap of differential accessible regions of enriched motifs}
\usage{
plotATACMotifHeatmap(
  objMOList,
  pValueAdj = 0.05,
  pValue = NULL,
  log2FEnrich = NULL
)
}
\arguments{
\item{objMOList}{An MOList object containing the differential accessible
regions}

\item{pValueAdj}{The cutoff for adjusted p-value, default is 0.05}

\item{log2FEnrich}{The cutoff for log2 fold enrichment, default is NULL}

\item{PValue}{The cutoff for p-value. If pValue is specified, pValueAdj
will be ignored}
}
\value{
A ComplexHeatmap object
}
\description{
This function generates a combined profile and heatmap of
             differential accessible regions of enriched motifs based on the
             ATACseq data. This is a wrapper function for the
             plotMotifHeatmaps function in the ChIPseeker package.
}
\details{
The motif enrichment analysis can be exceptionally stringent for
         ATACseq data. The users can specify the cutoff for the adjusted
         p-value or the regular p-value. The users can also specify the
         log2 enrichment fold change cutoff. Please use this function to
         explore the enriched motifs in the differential accessible regions
         and decide the best combination of cutoffs for constructing the
         transcriptional regulatory network. If pValue is specified,
         pValueAdj will be ignored.
}
\examples{
# Suppose that we have an MOList object called objMOList, which contains the
# differential accessible regions as an element named DEATAC, with motif
# enrichment analysis results

# Example 1: Plot the motif heatmap by default parameters
\dontrun{
plotATACMotifHeatmap(objMOList)
}

# Example 2: Plot the motif heatmap with unadjusted p-value cutoff of 0.01
\dontrun{
plotATACMotifHeatmap(objMOList, pValue = 0.01)
}

# Example 3: Plot the motif heatmap with log2 fold enrichment cutoff of 1
\dontrun{
plotATACMotifHeatmap(objMOList, pValue = 0.01, log2FEnrich = 1)
}

}
