% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEOmics.R
\name{plotATACProfileHeatmap}
\alias{plotATACProfileHeatmap}
\title{Plotting the combined profile and heatmap of annotated peaks}
\usage{
plotATACProfileHeatmap(
  objMOList,
  upstream = 1000,
  downstream = 1000,
  by = "gene",
  type = "start_site"
)
}
\arguments{
\item{objMOList}{An MOList object containing the annotated peaks}

\item{upstream}{The upstream distance from the TSS, default is 1000}

\item{downstream}{The downstream distance from the TSS, default is 1000}

\item{by}{The feature to be used for grouping, default is "gene"}

\item{type}{The type of the feature, default is "start_site"}
}
\value{
A ggplot object
}
\description{
This function generates a combined profile and heatmap of
             annotated peaks based on the ATACseq data. This is a wrapper
             function for the peak_Profile_Heatmap function in the
             ChIPseeker package.
}
\details{
See ?ChIPseeker::peak_Profile_Heatmap for details on the values of
         the parameters.
}
\examples{
# Use the package-provided example data
data(expMOList, package = "IntegraTRN")

# Plotting the profile and heatmap
\dontrun{
plotATACProfileHeatmap(expMOList)
}

}
\references{
\insertRef{yu2015chipseeker}{IntegraTRN}

\insertRef{lawrence2013software}{IntegraTRN}
}
