% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRNet.R
\name{plotNetwork}
\alias{plotNetwork}
\alias{plotNetwork,TRNet-method}
\title{Plot TRNet object}
\usage{
plotNetwork(trn, interactive = FALSE, ...)
}
\arguments{
\item{trn}{A TRNet object}

\item{interactive}{A logical value indicating whether the plot should be
interactive. If TRUE, an interactive plot will be
plotted using the networkD3 package. If FALSE, a static
plot will be plotted using the igraph package.}

\item{...}{Additional arguments passed to the plot function for plotting
a static plot using the igraph package, or the forceNetwork
function for interactive plots using the networkD3. See the
respective package documentation for details.}
}
\description{
This function plots the TRNet object
}
\examples{
# Define some example edges
edges <- data.frame(
  regulator = c("A", "B", "C"),
  target = c("D", "E", "F"),
  regulatorType = c("miRNA", "TF", "TF")
)

# Create TRNet object
trn <- TRNet(edges, FALSE, "RNA-seq")

# Plot TRNet object
plotNetwork(trn)

# Plot interactive TRNet object
plotNetwork(trn, interactive = TRUE)

}
\references{
\insertRef{csardi2006igraph}{IntegraTRN}

\insertRef{networkd3}{IntegraTRN}
}
