% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEOmics.R
\name{plotSmallRNAPCAs}
\alias{plotSmallRNAPCAs}
\title{Plotting results for Principle Component Analysis (PCA) for each type of
small RNA}
\usage{
plotSmallRNAPCAs(objMOList, batch = NULL, col1 = "#440154", col2 = "#fde725")
}
\arguments{
\item{objMOList}{An MOList object containing the raw expression matrix for
small RNAs}

\item{batch}{A vector of factors for batch effect, default is NULL. Used to
correct batch effect in the PCA analysis for count normalization}

\item{col1}{The color for the first group, default is "#440154"}

\item{col2}{The color for the second group, default is "#fde725". If
continuous grouping variable is used, the color gradient will be
used instead of the two colors}
}
\value{
A list of ggplot objects
}
\description{
This function generates a PCA plot for each type of small RNA
             based on the normalized expression matrix. This function is
             used to explore the contribution of each type of small RNA to
             separating samples based on their phenotypes.
}
\examples{
# Use the package-provided example data
data(expMOList, package = "IntegraTRN")

# Generate a list of PCA plots for each type of small RNA
\dontrun{
pcaPlotList <- plotSmallRNAPCAs(expMOList)
}

}
\references{
\insertRef{villanueva2019ggplot2}{IntegraTRN}

\insertRef{ggrepel}{IntegraTRN}

\insertRef{dplyr}{IntegraTRN}

\insertRef{love2014moderated}{IntegraTRN}
}
