% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEOmics.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Volcano plot for visualizing differentially expressed mRNA}
\usage{
plotVolcano(
  objMOList,
  omic = c("RNAseq", "smallRNAseq", "proteomics"),
  log2FC = 0,
  adjP = 0.05,
  upColor = "firebrick3",
  downColor = "dodgerblue3",
  title = NULL,
  highlight = NULL,
  pointSize = 1
)
}
\arguments{
\item{objMOList}{An MOList object containing the differential expression
results for mRNA.}

\item{omic}{The omic type, one of "RNAseq", "smallRNAseq", or "proteomics".
Must match the cases as defined here. Default is "RNAseq".}

\item{log2FC}{A numeric value specifying the cutoff for log2 fold change.
Must be a non-negative number. Default is 0.}

\item{adjP}{A numeric value specifying the cutoff for adjusted p-value.
Must be between 0 and 1. Default is 0.05.}

\item{upColor}{A character string specifying the color for up-regulated
genes. Default is "firebrick3".}

\item{downColor}{A character string specifying the color for down-regulated
genes. Default is "dodgerblue3".}

\item{title}{A single character string specifying the title for the plot.
Default is NULL, which means no title is added.}

\item{highlight}{A character vector specifying the gene names to highlight
in the plot. The names of the genes must be consistent with
the row names of the provided raw data. Otherwise,
non-matching gene names will be ignored. Default is NULL.}

\item{pointSize}{A positive numeric value specifying the size of the points
in the plot. Default is 1.}
}
\value{
A ggplot object
}
\description{
This function generates a volcano plot for visualizing
             differentially expressed mRNA based on log2 fold change and
             adjusted p-value. The user can specify the cutoffs for log2 fold
             change and adjusted p-value. The user can also specify a list of
             genes to highlight in the plot. The colors for up- and down-
             regulated genes can be specified by the user, but a default
             color scheme is provided.
}
\examples{
# Use the package-provided example data
data("expMOList")

# Example 1: Generate the volcano plot by default parameters
plotVolcano(expMOList)

# Example 2: Generate the volcano plot with custom parameters
plotVolcano(expMOList,
  omic = "RNAseq",
  log2FC = 0,
  adjP = 0.01,
  upColor = "purple",
  downColor = "green",
  title = "Volcano plot for DE mRNA"
)

}
\references{
\insertRef{villanueva2019ggplot2}{IntegraTRN}

\insertRef{dplyr}{IntegraTRN}
}
