% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEOmics.R
\name{plotVolcanoSmallRNA}
\alias{plotVolcanoSmallRNA}
\title{Volcano plot for visualizing differentially expressed small RNAs}
\usage{
plotVolcanoSmallRNA(
  objMOList,
  log2FC = 1,
  adjP = 0.05,
  geneList = NULL,
  colScheme = "BuPu",
  title = NULL
)
}
\arguments{
\item{objMOList}{An MOList object containing the differential expression
results for small RNAs.}

\item{log2FC}{The cutoff for log2 fold change. Default is 1.}

\item{adjP}{The cutoff for adjusted p-value. Default is 0.05.}

\item{geneList}{A vector of genes to highlight in the plot.}

\item{colScheme}{A RColorBrewer color scheme for color-coding each type of
small RNAs. Default is "BuPu".}

\item{title}{The title for the plot. Default is NULL.}
}
\value{
A ggplot object
}
\description{
This function generates a volcano plot for visualizing
             differentially expressed small RNAs based on log2 fold change
             and adjusted p-value. The user can specify the cutoffs for log2
             fold change and adjusted p-value. The user can also specify a
             l ist of genes to highlight in the plot. The users can define
             a color scheme used for color-coding each type of small RNAs.
             The default color scheme is provided.
}
\examples{
# Suppose that we have an MOList object called objMOList, which contains the
# differential expression results for small RNAs as an element named
DESmallRNAseq

# Example 1: Generate the volcano plot by default parameters
\dontrun{
plotVolcanoSmallRNA(objMOList)
}

# Example 2: Generate the volcano plot with custom parameters
\dontrun{
plotVolcanoSmallRNA(objMOList,
  log2FC = 2,
  adjP = 0.01,
  geneList = c("hsa-miR-1-3p", "hsa-miR-2-3p"),
  colScheme = "magma",
  title = "Volcano plot for DE small RNA"
)
}

}
