% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRNAsmallRNAcoExprEstimate.R
\name{predictSmallRNAmRNAcoExpr}
\alias{predictSmallRNAmRNAcoExpr}
\title{Predict small RNA - mRNA interactions via tree-based ensemble learning}
\usage{
predictSmallRNAmRNAcoExpr(
  mRNATopTag,
  smallRNATopTag,
  smallRNATypes = "all",
  annoSncRNA,
  matchingRNAsmallRNA,
  ntree = 1000,
  nthreads = 1,
  treeMethod = "RF",
  seed = 91711
)
}
\arguments{
\item{mRNATopTag}{A TOPTag object containing the top mRNAs that are
differentially expressed cross the conditions.}

\item{smallRNATopTag}{A TOPTag object containing the top small RNAs that are
differentially expressed cross the conditions.}

\item{smallRNATypes}{A character vector containing the types of small RNAs to
be used for prediction. The types should be a valid type
defined during the annotation process. Must be a subset
of the types: "miRNA", "piRNA", "tRNA", "snRNA",
"snoRNA", "circRNA", or a single string "all" to use all
types.}

\item{annoSncRNA}{A list containing the transcripts of the small RNAs
in each respective type. The list should be named
consistently as defined in annotateSmallRNA().}

\item{matchingRNAsmallRNA}{A list of two numeric vectors with equal length,
where each element in the first vector is the
index of the RNAseq sample, and the corresponding
element in the second vector is the index of the
matched small RNAseq sample. These indices match
to the sample names in the RNAseqSamples and
SmallRNAseqSamples slots of the MOList object.}

\item{ntree}{The number of trees to be used in the ensemble learning.}

\item{nthreads}{The number of threads to be used in the ensemble learning.}

\item{treeMethod}{The tree method to be used in the ensemble learning, either
"RF" for random forest or "ET" for extra trees. See the
documentation of the GENIE3 package for details.}

\item{seed}{The seed to be used in the ensemble learning for reproducibility.}
}
\value{
A data frame containing the predicted interactions, with the
        following format:
\itemize{
\item{regulator}{The names of the small RNAs that are predicted to regulate
                 the target genes.}
\item{target}{The names of the target genes that are predicted to be
              regulated by the small RNAs.}
\item{weight}{The weight of the predicted interactions.}
}
}
\description{
This function predicts small NC RNA - mRNA interactions via
             tree-based ensemble learning using the GENIE3 algorithm.
             The prediction is directed from small RNA to mRNA.
}
\note{
The final result includes all predicted interactions, without any
      filtering on the weight of the predicted interactions.
}
\keyword{internal}
