% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processPeaks.R
\name{processPeakOverlap}
\alias{processPeakOverlap}
\title{Processing user input peak files}
\usage{
processPeakOverlap(objMOList)
}
\arguments{
\item{objMOList}{An object of class MOList}
}
\value{
An object of class MOList, with a DETag added containing a master
        peak set that annotates where the peaks come from
}
\description{
This function processes user input peak files to check for
             overlaps in genomic coordinates. Although this process is not
             necessary for high-quality peak files, it improves the
             robustness of the subsequent motif enrichment analysis and
             peak annotation. Any peaks that are found to be overlapping
             with a minimum of 1 bp overlap will be merged into a single
             peak, for peaks from a single condition. Since the two input
             peak files that already represents differential accessible
             regions, we do not merge peaks across conditions.
}
\keyword{internal}
