% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{protein_heart}
\alias{protein_heart}
\title{Protein abundance in fetal heart tissues}
\format{
A data frame with 489 rows (proteins) and 6 columns (samples)
\describe{
 \item{sample_1, sample_2, sample_3, sample_4, sample_5, sample_6}{Protein
 abundance values for each sample}
}
}
\source{
Paul Delgado Olguin Lab, The Hospital for Sick Children, Toronto,
ON, Canada
}
\usage{
protein_heart
}
\description{
A proteomics experiment conducted on fetal ventricular heart tissues from
6 individuals, using LC-MS/MS. Only concensus proteins found in all samples
in the same conditions are included.
}
\examples{
\dontrun{
protein_heart
}

}
\keyword{datasets}
