% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{protein_heart_samples}
\alias{protein_heart_samples}
\title{Sample information for proteomics of heart tissues}
\format{
A data frame with 6 rows (samples) and 3 columns (fields)
\describe{
 \item{Sample}{Sample name corresponding to the proteomics gene expression
 dataset}
 \item{Age}{Age of the fetus in gestational weeks}
}
}
\source{
Paul Delgado Olguin Lab, The Hospital for Sick Children, Toronto,
ON, Canada
}
\usage{
protein_heart_samples
}
\description{
A selected set of sample information for the proteomics experiment conducted
on fetal ventricular heart tissues from 6 individuals
}
\examples{
\dontrun{
protein_heart_samples
}
}
\keyword{datasets}
