% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proteinGeneIDConvert}
\alias{proteinGeneIDConvert}
\title{Protein to gene mapping}
\format{
A data frame with 1421 rows (proteins) and 2 columns (fields)
\describe{
\item{protein}{Protein names}
\item{gene}{Gene names}
}
}
\usage{
proteinGeneIDConvert
}
\description{
A data frame containing the mapping between protein names and gene names
used in the example proteomics and RNAseq datasets. This is only a subset
of the full mapping and does not cover all proteins in the example data for
illustration purposes.
}
\examples{
\dontrun{
proteinGeneIDConvert
}

}
\references{
\insertRef{durinck2009mapping}{IntegraTRN}
}
\keyword{datasets}
