% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOList.R
\name{proteomics,MOList-method}
\alias{proteomics,MOList-method}
\alias{proteomics}
\title{Getter for the proteomics data from the MOList object}
\usage{
\S4method{proteomics}{MOList}(x)
}
\arguments{
\item{x}{An object of class MOList for retrieving omics data}
}
\value{
A numeric matrix containing the proteomics data
}
\description{
Getter for the proteomics data from the MOList object
}
\examples{
\dontrun{
# Using the example MOList object
dataProteomics <- proteomics(myMOList)
}

}
