% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOList.R
\name{RNAseq,MOList-method}
\alias{RNAseq,MOList-method}
\alias{RNAseq}
\title{Getter for the RNAseq data from the MOList object}
\usage{
\S4method{RNAseq}{MOList}(x)
}
\arguments{
\item{x}{An object of class MOList for retrieving omics data}
}
\value{
A numeric matrix containing the RNAseq data
}
\description{
Getter for the RNAseq data from the MOList object
}
\examples{
\dontrun{
# Using the example MOList object
dataRNAseq <- RNAseq(myMOList)
}

}
