% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RNAseq_heart}
\alias{RNAseq_heart}
\title{Gene expression of fetal heart tissues}
\format{
A data frame with 13043 rows (genes) and 54 columns (field, samples)
\describe{
 \item{gene}{Gene name}
 \item{sample_1, ..., sample_53}{Gene expression values for each sample}
}
}
\source{
Paul Delgado Olguin Lab, The Hospital for Sick Children, Toronto,
ON, Canada
}
\usage{
RNAseq_heart
}
\description{
An RNAseq experiment conducted on fetal ventricular heart tissues from 53 
individuals
}
\examples{
\dontrun{
 RNAseq_heart
}

}
\references{
Ahmed, A., Liang, M., Chi, L., Zhou, Y. Q., Sled, J. G., Wilson, M. D., & 
Delgado-Olguín, P. (2021). Maternal obesity persistently alters cardiac 
progenitor gene expression and programs adult-onset heart disease 
susceptibility. Molecular Metabolism, 43. 
https://doi.org/10.1016/j.molmet.2020.101116
}
\keyword{datasets}
