% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RNAseq_heart_samples}
\alias{RNAseq_heart_samples}
\title{Sample information for RNAseq of heart tissues}
\format{
A data frame with 53 rows (samples) and 4 columns (fields)
\describe{
 \item{Sample}{Sample name corresponding to the RNAseq gene expression
 dataset}
 \item{Age}{Age of the fetus in gestational weeks}
 \item{Sex}{Sex of the fetus}
 \item{Batch}{Batch number of the RNAseq experiment, represented from
 A to F}
}
}
\source{
Paul Delgado Olguin Lab, The Hospital for Sick Children, Toronto,
ON, Canada
}
\usage{
RNAseq_heart_samples
}
\description{
A selected set of sample information for the RNA-seq experiment conducted on
fetal ventricular heart tissues from 53 individuals
}
\examples{
\dontrun{
RNAseq_heart_samples
}

}
\references{
Ahmed, A., Liang, M., Chi, L., Zhou, Y. Q., Sled, J. G., Wilson, M. D., &
Delgado-Olguín, P. (2021). Maternal obesity persistently alters cardiac
progenitor gene expression and programs adult-onset heart disease
susceptibility. Molecular Metabolism, 43.
https://doi.org/10.1016/j.molmet.2020.101116
}
\keyword{datasets}
