% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRNAsmallRNAcoExprEstimate.R
\name{runGENIE3}
\alias{runGENIE3}
\title{Run GENIE3 for predicted interactions}
\usage{
runGENIE3(
  exprMatrix,
  regulators,
  ntree = 1000,
  nthreads = 1,
  treeMethod = "RF",
  seed = 91711
)
}
\arguments{
\item{exprMatrix}{A numeric matrix containing the expression data.}

\item{regulators}{A character vector containing the names of the regulators.}

\item{ntree}{The number of trees to be used in the ensemble learning.}

\item{nthreads}{The number of threads to be used in the ensemble learning.}

\item{treeMethod}{The tree method to be used in the ensemble learning, either
"RF" for random forest or "ET" for extra trees. See the
documentation of the GENIE3 package for details.}

\item{seed}{The seed to be used in the ensemble learning for reproducibility.}
}
\value{
A list containing the predicted interactions, with the
       following format:
\itemize{
\item{regulatoryGene}{The names of the small RNAs that are predicted to regulate
                      the target genes.}
\item{targetGene}{The names of the target genes that are predicted to be
                  regulated by the small RNAs.}
\item{weight}{The weight of the predicted interactions.}
}
}
\description{
This function runs GENIE3 for predicted interactions.
}
\keyword{internal}
