% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runIntegraTRN.R
\name{runIntegraTRN}
\alias{runIntegraTRN}
\title{Launch Shiny App for package IntegraTRN}
\usage{
runIntegraTRN()
}
\value{
None, but opens the Shiny app
}
\description{
This is a function that launches the Shiny app for the package
             IntegraTRN. The Shiny app permits the users to perform
             differential analysis of the user-input omics data as well as
             the visualization of the results. It also further allows the
             users to perform the network construction and visualization of
             the network based on input and inferred regulatory interactions.
}
\examples{
\dontrun{
runIntegraTRN()
}

}
\references{
\insertRef{csardi2006igraph}{IntegraTRN}

\insertRef{huynh2010inferring}{IntegraTRN}

\insertRef{chang2020mirnet}{IntegraTRN}
}
