% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEOmics.R
\name{selectedMotifs}
\alias{selectedMotifs}
\title{Select enriched motifs}
\usage{
selectedMotifs(
  enrichedMotifs,
  pValueAdj = 0.05,
  pValue = NULL,
  log2FEnrich = NULL
)
}
\arguments{
\item{enrichedMotifs}{A SummarizedExperiment object containing the motif
enrichment results}

\item{pValueAdj}{The cutoff for adjusted p-value, default is 0.05}

\item{pValue}{The cutoff for p-value, default is NULL. if pValue is
specified, pValueAdj will be ignored}

\item{log2FEnrich}{The cutoff for log2 fold enrichment, default is NULL}
}
\value{
A logical vector with same number of rows as the number of motifs
        in the input object
}
\description{
This function selects enriched motifs based on user-specified
             cutoff and the motif enrichment result. If pValue is specified,
             pValueAdj will be ignored.
}
\keyword{internal}
