% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructNetwork.R
\name{setOmicCutoffs}
\alias{setOmicCutoffs}
\title{Setting cutoffs for the omics data}
\usage{
setOmicCutoffs(
  rnaAdjPval = 0.05,
  rnaLogFC = 0,
  rnaTopGenes = 1,
  smallRNAAdjPval = 0.05,
  smallRNALogFC = 0,
  smallRNATopGenes = 1,
  proteomicsAdjPval = 0.05,
  proteomicsLogFC = 0,
  atacMotifAdjPval = 0.05,
  atacMotifPval = NULL,
  atacMotifLogFC = NULL
)
}
\arguments{
\item{rnaAdjPval}{The adjusted p-value cutoff for the RNAseq data, must be
a numeric value between 0 and 1. Default is 0.05.}

\item{rnaLogFC}{The log fold change cutoff for the RNAseq data, must be a
positive number. Default is 0.}

\item{rnaTopGenes}{A numeric value indicating either the fraction (0-1) of
top differential genes or the number (1-Inf) of top
differential genes. If the number specified is greater
than the number of DE genes based on logFCCutoff and
pCutoff, then topGenes will be set to the number of DE
genes. This is to select the top differential mRNAs.
To include all the differential mRNAs, set this to 1.
Default is 1.}

\item{smallRNAAdjPval}{The adjusted p-value cutoff for the smallRNAseq data.
Must be a numeric value between 0 and 1. Default is
0.05.}

\item{smallRNALogFC}{The log fold change cutoff for the smallRNAseq data.
Must be a positive number. Default is 0.}

\item{smallRNATopGenes}{A numeric value indicating either the fraction (0-1)
of top differential genes or the number (1-Inf) of
top differential genes. If the number specified is
greater than the number of DE genes based on
logFCCutoff and pCutoff, then topGenes will be set
to the number of DE genes. This is to select the top
differential small RNAs. To include all the
differential small RNAs, set this to 1. Default is
1.}

\item{proteomicsAdjPval}{The adjusted p-value cutoff for the proteomics data.
Must be a numeric value between 0 and 1. Default is
0.05.}

\item{proteomicsLogFC}{The log fold change cutoff for the proteomics data.
Must be a positive number. Default is 0.}

\item{atacMotifAdjPval}{The adjusted p-value cutoff for the ATACseq motif
enrichment analysis. Must be a numeric value between
0 and 1. Default is 0.05.}

\item{atacMotifPval}{The p-value cutoff for the ATACseq motif enrichment
analysis. This value is defaulted to NULL, but if the
user specifies this value, then the adjusted p-value
cutoff will be ignored.}

\item{atacMotifLogFC}{The log2 fold enrichment cutoff for the ATACseq motif
enrichment analysis.}
}
\value{
An OMICutoffs object containing the cutoffs for the omics data. This
        is essentially a list with defined elements. Once setOmicCutoffs is
        called, users can freely access or modify the cutoffs by using the
        $ operator. See the examples for details.
}
\description{
This function sets the cutoffs for the omics data. The cutoffs
             are used to determine the key differentially regulated/expressed
             genes/miRNAs/TFs. The users can set the cutoffs based on their
             interpretation during the exploratory analysis on the
             differential analysis process. Here, the cutoffs are set to
             define a set of genes/miRNAs/TFs that are the key to such
             differential regulation/expression, and only these genes/miRNAs/
             TFs will be used to construct the network. Any values set on
             data that is not available will be ignored.
}
\examples{
# Example 1: Set the cutoffs for the omics data
omiCutoffs <- setOmicCutoffs(
  rnaAdjPval = 0.05,
  rnaLogFC = 2,
  rnaTopGenes = 0.1, # Top 10\% of DE genes
  smallRNAAdjPval = 0.05,
  smallRNALogFC = 2,
  smallRNATopGenes = 200, # Top 200 DE small RNAs
  proteomicsAdjPval = 0.05,
  proteomicsLogFC = 1
)

# Example 2: Access the cutoffs for the RNAseq data
omiCutoffs$rnaAdjPval
omiCutoffs$rnaLogFC

# Example 3: Modify the cutoffs for the proteomics data
omiCutoffs$proteomicsAdjPval <- 0.01
omiCutoffs$proteomicsLogFC <- 1.5

}
