% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOList.R
\name{setOmics}
\alias{setOmics}
\title{Setting non-mRNA omics data to MOList object}
\usage{
setOmics(
  objMOList,
  smallRNAseq,
  smallRNAGroupBy,
  proteomics,
  proteomicsGroupBy,
  peakCond1,
  peakCond2
)
}
\arguments{
\item{objMOList}{An object of class MOList for appending/exchanging omics
data}

\item{smallRNAseq}{A numeric matrix containing the smallRNAseq data}

\item{smallRNAGroupBy}{A vector of grouping information for the smallRNAseq
data}

\item{proteomics}{A numeric matrix containing the proteomics data}

\item{proteomicsGroupBy}{A vector of grouping information for the proteomics
data}

\item{peakCond1}{A data frame containing the differentially accessible ATAC
peaks for condition 1}

\item{peakCond2}{A data frame containing the differentially accessible ATAC
peaks for condition 2}
}
\value{
An object of class MOList
\itemize{
 \item \code{RNAseq}: A numeric matrix containing the RNAseq data
 \item \code{RNAseqSamples}: A list containing the sample names and grouping
                           information for the RNAseq data
 \item \code{smallRNAseq}: A numeric matrix containing the smallRNAseq data
 \item \code{smallRNAseqSamples}: A list containing the sample names and
                              grouping information for the smallRNAseq data
 \item \code{proteomics}: A numeric matrix containing the proteomics data
 \item \code{proteomicsSamples}: A list containing the sample names and
                            grouping information for the proteomics data
 \item \code{ATACpeaks}: A list containing the differentially accessible ATAC
                        peaks for condition 1 and condition 2
}
}
\description{
This function sets the non-mRNA omics data to the MOList object.
}
\keyword{internal}
