% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DETag.R
\name{show,DETag-method}
\alias{show,DETag-method}
\alias{print,DETag-method}
\title{Print DETag}
\usage{
\S4method{show}{DETag}(object)
}
\arguments{
\item{object}{An object of the DETag class}
}
\description{
This function prints the DETag object
}
\examples{
# Create an example data frame
deResult <- data.frame(
 log2FoldChange = runif(10),
 padj = runif(10),
 pvalue = runif(10)
)
rownames(deResult) <- paste0("gene", seq_len(10))

# Create a normalized counts matrix
normalizedCounts <- matrix(runif(100), nrow = 10)

# Create an object of the DETag class
deTag <- DETag(deResult, "DESeq2", normalizedCounts)

# Print the object
print(deTag)

# or simply do
deTag

}
