% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOList.R
\name{show,MOList-method}
\alias{show,MOList-method}
\alias{print,MOList-method}
\title{print MOList object}
\usage{
\S4method{show}{MOList}(object)
}
\arguments{
\item{object}{An object of the DETag class}
}
\description{
This function prints the MOList object
}
\examples{
# Use the package-provided example MOList object
data("expMOList")

# Print the MOList object
print(expMOList)

# or simply type the object name
expMOList

}
