% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEAKTag.R
\name{show,PEAKTag-method}
\alias{show,PEAKTag-method}
\title{print PEAKTag}
\usage{
\S4method{show}{PEAKTag}(object)
}
\arguments{
\item{object}{A PEAKTag object}
}
\description{
The print method for the PEAKTag class.
}
\examples{
# Use the package's provided example data
data(expMOList)

# Print the object
print(expMOList$DEATAC)

# Or simply type the object name
expMOList$DEATAC

}
