% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOPTag.R
\name{show,TOPTag-method}
\alias{show,TOPTag-method}
\alias{print,TOPTag-method}
\title{print TOPTag}
\usage{
\S4method{show}{TOPTag}(object)
}
\arguments{
\item{object}{A TOPTag object.}
}
\description{
The print method for the TOPTag class.
}
\examples{
# Use the package provided example data
data("expMOList")

# Extract an DETag object from RNA-seq data
deTag <- expMOList$DERNAseq

# Create a TOPTag object
topTag <- TOPTag(deTag, topGenes = 0.2)

# Print the TOPTag object
print(topTag)

# or simply type the object name
topTag

}
