% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRNet.R
\name{show,TRNet-method}
\alias{show,TRNet-method}
\title{show method for TRNet object}
\usage{
\S4method{show}{TRNet}(object)
}
\arguments{
\item{object}{A TRNet object}
}
\description{
This function prints the TRNet object
}
\examples{
# Define some example edges
edges <- data.frame(
  regulator = c("A", "B", "C"),
  target = c("D", "E", "F"),
  regulatorType = c("miRNA", "TF", "TF")
)

# Create TRNet object
trn <- TRNet(edges, FALSE, "RNA-seq")

# Print TRNet object
trn

}
