% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOList.R
\name{smallRNAseq,MOList-method}
\alias{smallRNAseq,MOList-method}
\alias{smallRNAseq}
\title{Getter for the smallRNAseq data from the MOList object}
\usage{
\S4method{smallRNAseq}{MOList}(x)
}
\arguments{
\item{x}{An object of class MOList for retrieving omics data}
}
\value{
A numeric matrix containing the smallRNAseq data
}
\description{
Getter for the smallRNAseq data from the MOList object
}
\examples{
# Generating some example data
# Note that the exact value could differ based on the random seed
RNAseq <- matrix(sample(1:100, 100, replace = TRUE), ncol = 10)
RNAGroupBy <- rep(c("A", "B"), each = 5)
smallRNAseq <- matrix(sample(1:100, 20, replace = TRUE), ncol = 4)
smallRNAGroupBy <- rep(c("A", "B"), each = 2)

# Constructing the MOList object
objMOList <- MOList(RNAseq = RNAseq,
                    RNAGroupBy = RNAGroupBy,
                    smallRNAseq = smallRNAseq,
                    smallRNAGroupBy = smallRNAGroupBy)

# Retrieving the smallRNAseq data
dataSmallRNAseq <- smallRNAseq(objMOList)

}
