% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smallRNAseq_heart}
\alias{smallRNAseq_heart}
\title{Expression of small RNA in fetal heart tissues}
\format{
A data frame with 14801 rows (small RNA transcripts) and 38 columns 
(field, samples)
\describe{
 \item{transcript}{Transcript name}
 \item{sample_1, ..., sample_37}{Transcript expression values for each
 sample}
}
}
\source{
Paul Delgado Olguin Lab, The Hospital for Sick Children, Toronto,
ON, Canada
}
\usage{
smallRNAseq_heart
}
\description{
A small RNAseq experiment conducted on fetal ventricular heart tissues from
37 individuals
}
\examples{
\dontrun{
smallRNA_heart
}
}
\keyword{datasets}
