% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smallRNAseq_heart_samples}
\alias{smallRNAseq_heart_samples}
\title{Sample information for small RNAseq of heart tissues}
\format{
A data frame with 37 rows (samples) and 3 columns (fields)
\describe{
 \item{Age}{Age of the fetus in gestational weeks}
 \item{Sex}{Sex of the fetus}
}
}
\source{
Paul Delgado Olguin Lab, The Hospital for Sick Children, Toronto,
ON, Canada
}
\usage{
smallRNAseq_heart_samples
}
\description{
A selected set of sample information for the small RNA-seq experiment
conducted on fetal ventricular heart tissues from 37 individuals
}
\examples{
\dontrun{
smallRNA_heart_samples
}

}
\references{
\insertRef{adar2023integrate}{IntegraTRN}
}
\keyword{datasets}
