% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processSmallRNA.R
\name{sncAnnoCoverage}
\alias{sncAnnoCoverage}
\title{Check annotation coverage}
\usage{
sncAnnoCoverage(transcripts, objMOList, anno)
}
\arguments{
\item{transcripts}{A character vector of transcript names}

\item{objMOList}{A object of class MOList}

\item{anno}{A character string indicating the annotation to be used. Either
"human" or "userAnno"}
}
\value{
A boolean indicating whether the annotation covers all transcript
}
\description{
This function checks if the annotation covers all transcript
             names provided
}
\references{
\insertRef{dplyr}{IntegraTRN}
}
\keyword{internal}
