% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sncRNAAnnotation}
\alias{sncRNAAnnotation}
\title{Small RNA type annotation for human}
\format{
A list of 6 character vectors of small RNA
\describe{
\item{miRNA}{A character vector of miRNA names}
\item{piRNA}{A character vector of piRNA names}
\item{snoRNA}{A character vector of snoRNA names}
\item{snRNA}{A character vector of snRNA names}
\item{tRNA}{A character vector of tRNA names}
\item{circRNA}{A character vector of circRNA names}
}
}
\source{
miRbase, piRNAbank, piRBase, GtRNAdb, circBase, GENCODE, and
        piRNACluster
}
\usage{
sncRNAAnnotation
}
\description{
A list of 6 character vectors of small RNA. Each element of the list
represents a small RNA type. The 6 types are: miRNA, piRNA, snoRNA, snRNA,
tRNA, and circRNA. The character vectors contain the names of small RNAs
belonging to the corresponding type, using the nomenclature from the
listed databases.
}
\examples{
\dontrun{
sncRNAAnnotation
}
}
\references{
\insertRef{li2020compsra}{IntegraTRN}

\insertRef{kozomara2010mirbase}{IntegraTRN}

\insertRef{chan2016gtrnadb}{IntegraTRN}

\insertRef{harrow2012gencode}{IntegraTRN}

\insertRef{glavzar2014circbase}{IntegraTRN}

\insertRef{sai2008pirnabank}{IntegraTRN}

\insertRef{zhang2014pirbase}{IntegraTRN}

\insertRef{rosenkranz2016pirna}{IntegraTRN}
}
\keyword{datasets}
