% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tf2Genes}
\alias{tf2Genes}
\title{Example TF - mRNA interactions}
\format{
A list of 2 vectors of the same length. Each representing a pair of
       TF-target interactions.
\describe{\
\item{regulator}{A vector of TFs}
\item{target}{A vector of upregulated fetal heart mRNAs}
}
}
\usage{
tf2Genes
}
\description{
A list of TF-target interactions on fetal heart differentially expressed
genes. The list contains 2 vectors of the same length. Each representing a
pair of TF-target interactions. Curated partially from randomly selected
TF-target interactions from TRRUST v2 using miRNet 2.0. Partially simulated
using simulated ATACseq peaks.
}
\examples{
\dontrun{
tf2Genes
}

}
\references{
\insertRef{chang2020mirnet}{IntegraTRN}
}
\keyword{datasets}
