% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOPTag.R
\name{TOPTag}
\alias{TOPTag}
\title{Constructor for TOPTag class}
\usage{
TOPTag(
  object,
  logFCCutoff = 1,
  pCutoff = 0.05,
  topGenes = 0.1,
  direction = "both"
)
}
\arguments{
\item{object}{A DETag object.}

\item{logFCCutoff}{A numeric value indicating the absolute log2 fold change
cutoff used to filter the differential expression results.}

\item{pCutoff}{A numeric value indicating the adjusted p-value cutoff used to
filter the differential expression results.}

\item{topGenes}{A numeric value indicating either the fraction (0-1) of top
differential genes or the number (1-Inf) of top differential
genes. If the number specified is greater than the number of
DE genes based on logFCCutoff and pCutoff, then topGenes will
be set to the number of DE genes. Default to select the top
10 percent of differential genes.}

\item{direction}{A character string indicating the direction of the
differential expression. Default to "both". Other options
include "up" and "down".}
}
\value{
A TOPTag object.
}
\description{
The constructor for the TOPTag class.
}
\examples{
# Use the package provided example data
data("expMOList")
# Extract an DETag object from RNA-seq data
deTag <- expMOList$DERNAseq

# Example 1: Select the top 20\% of differential genes with default cutoffs
topTag <- TOPTag(deTag, topGenes = 0.2)
topTag

# Example 2: Select the top 50 differential genes with default cutoffs
\dontrun{
# In this example, a warning should be expected. This is because 50 is
# greater than the number of differentially expressed genes in the example
# data
topTag <- TOPTag(deTag, topGenes = 50)
topTag
}

# Example 3: Select the top 20\% of differential genes with custom cutoffs
topTag <- TOPTag(deTag, logFCCutoff = 0, pCutoff = 0.01, topGenes = 0.2)
topTag

# Example 4: Select the top 20\% of up-regulated genes with default cutoffs
topTag <- TOPTag(deTag, topGenes = 0.2, direction = "up")
topTag

}
\references{
\insertRef{dplyr}{IntegraTRN}

\insertRef{reimand2019pathway}{IntegraTRN}
}
