% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{upregmiR2gene}
\alias{upregmiR2gene}
\title{miRNet curated miRNA-target interactions on upregulated fetal heart mRNAs}
\format{
A list of 2 vectors of the same length. Each representing a pair of
        miRNA-target interactions.
\describe{\
\item{regulator}{A vector of miRNAs}
\item{target}{A vector of upregulated fetal heart mRNAs}
}
}
\source{
Paul Delgado Olguin Lab, The Hospital for Sick Children, Toronto,
ON, Canada
}
\usage{
upregmiR2gene
}
\description{
A list of miRNA-target interactions on upregulated fetal heart mRNAs
curated from miRNet
}
\examples{
\dontrun{
upregmiR2gene
}

}
\references{
Le Chang, Guangyan Zhou, Othman Soufan, Jianguo Xia, miRNet 2.0:
            network-based visual analytics for miRNA functional analysis and
            systems biology, Nucleic Acids Research, Volume 48, Issue W1, 02
            July 2020, Pages W244–W251, https://doi.org/10.1093/nar/gkaa467
}
\keyword{datasets}
