% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffOmicsAnalysis.R
\name{validateDataAnno}
\alias{validateDataAnno}
\title{Validate the input data and annotations on the samples}
\usage{
validateDataAnno(objMOList, annoList)
}
\arguments{
\item{objMOList}{A MOList object containing the omics data}

\item{annoList}{A list containing the annotations on the samples of the omics
                data, must contain fields RNAseq, smallRNAseq, and proteomics
\itemize{
 \item RNAseq: A character/numeric vector for annotating each sample in the
               RNAseq data, must be the same length as the number of samples
               in the RNAseq data
\item smallRNAseq: A character/numeric vector for annotating each sample in
                   the small RNAseq data, must be the same length as the
                   number of samples in the small RNAseq data
\item proteomics: A character/numeric vector for annotating each sample in
                  the protein data, must be the same length as the number of
                  samples in the protein data
}}
}
\description{
This function validates the input data and annotations on the
             samples. The input data must be a MOList object, and the
             annotations on the samples must be a list containing the
             annotations on the RNAseq, small RNAseq, and protein data, in
             the order of RNAseq, small RNAseq, and protein.
}
\keyword{internal}
