% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DETag.R
\name{validateDETagSlots}
\alias{validateDETagSlots}
\title{Validator of the DETag class slots}
\usage{
validateDETagSlots(DEResult, method)
}
\arguments{
\item{DEResult}{A data frame containing the differential expression analysis
results}

\item{method}{A character string specifying the program used for the
analysis, must be one of the valid METHODS defined}
}
\description{
This function validates the DETag class
}
\examples{
# Create an example data frame
deResult <- data.frame(
  gene = paste0("gene", seq_len(10)),
  logFC = runif(10),
  adj.P.Val = runif(10)
)

# Create an object of the DETag class
deTag <- DETag(deResult, "DESeq2")

# Validate the object
validateDETagSlots(deTag)

}
\keyword{internal}
