% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructNetwork.R
\name{validateInteractionAdjList}
\alias{validateInteractionAdjList}
\title{Validate external interaction data}
\usage{
validateInteractionAdjList(adjList)
}
\arguments{
\item{adjList}{A list of interactions in adjacent list format}
}
\value{
A list of validated interactions in adjacent list format
}
\description{
This function validates the external interaction data provided
             by the user and returns a list of validated data.
}
\keyword{internal}
