% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOList.R
\name{validateMatrix}
\alias{validateMatrix}
\title{Validate matrix and grouping information}
\usage{
validateMatrix(matrix, groupBy)
}
\arguments{
\item{matrix}{A numeric matrix containing the sequencing data}

\item{groupBy}{A vector of grouping information for the sequencing data}
}
\description{
This function validates the matrix and grouping information for
             the sequencing data. The matrix must be a numeric matrix, and
             the grouping information must be a vector of the same length as
             the number of samples in the sequencing data.
}
\examples{
\dontrun{
# Assuming RNAseq, smallRNAseq, and proteomics are matrices of count data

# Validating the RNAseq data
validateMatrix(RNAseq, RNAGroupBy)
# Validating the smallRNAseq data
validateMatrix(smallRNAseq, smallRNAGroupBy)
# Validating the proteomics data
validateMatrix(proteomics, proteomicsGroupBy)
}

}
\keyword{internal}
