% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOList.R
\name{validateMOInputs}
\alias{validateMOInputs}
\title{Validate the inputs for the MOList constructor}
\usage{
validateMOInputs(
  RNAseq,
  RNAGroupBy,
  smallRNAseq,
  smallRNAGroupBy,
  proteomics,
  proteomicsGroupBy,
  peakCond1,
  peakCond2
)
}
\arguments{
\item{RNAseq}{A numeric matrix containing the RNAseq data}

\item{RNAGroupBy}{A vector of grouping information for the RNAseq data}

\item{smallRNAseq}{A numeric matrix containing the smallRNAseq data}

\item{smallRNAGroupBy}{A vector of grouping information for the smallRNAseq
data}

\item{proteomics}{A numeric matrix containing the proteomics data}

\item{proteomicsGroupBy}{A vector of grouping information for the proteomics
data}

\item{peakCond1}{A data frame containing the ATAC peaks for condition 1}

\item{peakCond2}{A data frame containing the ATAC peaks for condition 2}
}
\description{
This function validates the inputs for the MOList constructor.
             It uses the validateMatrix function to validate the sequencing
             data, and checks for missing values in the chromosome
             information in the ATAC peaks.
}
\examples{
\dontrun{
# Assuming RNAseq, smallRNAseq, and proteomics are matrices of count data

# Validating the inputs for the MOList constructor
validateMOInputs(
  RNAseq, RNAGroupBy, smallRNAseq, smallRNAGroupBy,
  proteomics, proteomicsGroupBy, peakCond1, peakCond2
)
}
}
\references{
Fischer, D. (2020). GenomicTools.fileHandler: File handlers for genomic data
analysis.
}
\keyword{internal}
