% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOList.R
\name{validateMOList}
\alias{validateMOList}
\title{Validator of the MOList S4 object}
\usage{
validateMOList(objMOList)
}
\arguments{
\item{objMOList}{An object of class MOList with any form of modification}
}
\description{
This function validates the MOList object. It provides sanity
             checks for the omics data and sample information specified in
             each slot of the MOList object. All S4 methods of this class
             should call this function to validate the MOList object after
             any modification is performed.
}
\keyword{internal}
