% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOList.R
\name{validatePeaks}
\alias{validatePeaks}
\title{Validate BED files for peak information}
\usage{
validatePeaks(peak)
}
\arguments{
\item{peak}{A data frame containing the peak information in the BED format}
}
\description{
This function validates the BED files for peak information.
             The BED files are expected to contain minimally the chromosome,
             start, and end coordinates of the peaks. This function provides
             extendibility for future implementation of additional data
             fields in the BED files, such as ChIP-seq peaks. Assumes that
             the input peak file is not NULL
}
\references{
\insertRef{grandi2022chromatin}{IntegraTRN}
}
\keyword{internal}
