% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRNAsmallRNAcoExprEstimate.R
\name{validateSampleDFs}
\alias{validateSampleDFs}
\title{Validate RNAseq and small RNAseq sample data frames for matching}
\usage{
validateSampleDFs(objMOList, sampleDFRNAseq, sampleDFSmallRNAseq)
}
\arguments{
\item{objMOList}{An object of class MOList}

\item{sampleDFRNAseq}{A data frame containing the RNAseq sample information.
Each sample is a row in the data frame, with row names
being the sample names that correspond to the sample
names in the MOList object.}

\item{sampleDFSmallRNAseq}{A data frame containing the small RNAseq sample
information. Each sample is a row in the data
frame, with row names being the sample names that
correspond to the sample names in the MOList
object.}
}
\description{
The data frames should contain the same number of samples
             and samples names as already documented in the MOList object.
             However, they do not need to be in the same order.
}
\keyword{internal}
