% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRNet.R
\name{writeTRN}
\alias{writeTRN}
\alias{writeTRN,TRNet-method}
\title{Exporting the TRN network to a third-party file format}
\usage{
writeTRN(trn, file, format = "graphml")
}
\arguments{
\item{trn}{A TRNet object}

\item{file}{A connection, or a single-length character string containing the
file to write to}

\item{format}{A character string specifying the format of the file to write
to. The default is "graphml". Supported formats include
"edgelist", "pajek", "ncol", "lgl", "graphml", "dimacs", "gml",
"dot", and "leda". See \code{\link[igraph]{write_graph}} for
details.}
}
\description{
This function exports the TRN network to a third-party file
             format. It is a wrapper function for the igraph::write_graph
             function.
}
\examples{
# Use the package data
data("expMOList")

# Set the cutoffs for the omics data
omiCutoffs <- setOmicCutoffs() # use default cutoffs

# Construct the network
\dontrun{
# Expect warnings that show conditional network construction steps
myTRNet <- constructTRN(expMOList, omiCutoffs, targetDirection = "up")
}

# Create a temporary file
tempFile <- tempfile(fileext = ".graphml")

# Export the network to the temporary file with default format
\dontrun{
writeTRN(myTRNet, file = tempFile)
}

# Or specify the file format
\dontrun{
tempFile2 <- tempfile(fileext = ".gml")
writeTRN(myTRNet, file = tempFile2, format = "gml")
}

}
\references{
\insertRef{csardi2006igraph}{IntegraTRN}
}
\seealso{
\code{\link[igraph]{write_graph}}
}
