% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEOmics.R
\name{annoExpr}
\alias{annoExpr}
\title{Annotate by expression}
\usage{
annoExpr(deg, log2FC, adjP)
}
\arguments{
\item{deg}{A data frame containing the differential expression analysis
results, must follow the described format in DETag-class}

\item{log2FC}{The cutoff for log2 fold change, a positive number}

\item{adjP}{The cutoff for adjusted p-value}
}
\description{
This function annotates whether a gene is up-regulated,
             down-regulated, or not differentially expressed
}
