% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processPeaks.R
\name{annotateATACPeaksMotif}
\alias{annotateATACPeaksMotif}
\title{Performing peak annotation and motif enrichment analysis}
\usage{
annotateATACPeaksMotif(
  objMOList,
  tssRegion = c(-3000, 3000),
  TxDb,
  annoDb = "org.Hs.eg.db",
  bsgenome,
  pwmL,
  fixedWidth = 500
)
}
\arguments{
\item{objMOList}{An object of class MOList}

\item{tssRegion}{The region around the TSS to annotate with
(default: +/- 3000)}

\item{TxDb}{The TxDb object to use for annotation, using one of the TxDb
packages}

\item{annoDb}{The annotation database to use for annotation, using one of
the org.*.eg.db packages. Must be a valid string for the name
of the package. For details, see the vignette of the ChIPseeker
package for custom annotation databases.}

\item{bsgenome}{The BSgenome object that contains the genome sequence
 Read the Bioconductor
vignette for the BSgenome package for more details.}

\item{pwmL}{A PWMatrixList object containing the position-weight matrices
(PWMs) to use for motif enrichment analysis. The PWMs can be
generated using the getMatrixSet() function from the TFBSTools
package. See the vignette of the TFBSTools package for more
details.}

\item{fixedWidth}{The fixed width to adjust the peak size for motif if
necessary (default: 500)}
}
\value{
An object of class MOList, with the ATACseq peaks annotated with
        genomic features and motif enrichment analysis results appended.
}
\description{
This function performs peak annotation and motif enrichment
             analysis on the ATACseq peaks. The peak annotation is performed
             using the ChIPseeker package, and the motif enrichment analysis
             is performed using the monaLisa package.
}
\details{
The annotation is performed using the TxDb object and annotation
         database specified by the user. Depending on the style of genomic
         coordinate representation (with or without the "chr" prefix for the
         chromosome name, or the use of "chrM", "M", "MT" for mitochondrial
         DNA), the TxDb object may need to be adjusted to match the style
         used in the peak file. See the vignette of the ChIPseeker package
         for more details. By default, we recommend using the UCSC style
         genomic coordinates, and the use of the
         TxDb.Hsapiens.UCSC.hg38.knownGene and org.Hs.eg.db packages for
         human samples, and the TxDb.Mmusculus.UCSC.mm10.knownGene and
         org.Mm.eg.db packages for mouse samples.
}
\examples{
# Suppose that myMOList is an object of class MOList with $DEATAC containing
# the ATACseq peaks as a DETag object
\dontrun{
# Generate position-weight matrices (PWMs) from the JASPAR database
pwmL <- TFBSTools::getMatrixSet(JASPAR2022::JASPAR2022,
  opts = list(
    matrixtype = "PWM",
    tax_group = "vertebrates"
  )
)
# Annotate the ATACseq peaks with genomic features and perform motif
# enrichment analysis
myMOList <- annotateATACPeaksMotif(myMOList,
  tssRegion = c(-3000, 3000),
  TxDb = TxDb.Hsapiens.UCSC.hg38.knownGene,
  annoDb = "org.Hs.eg.db",
  bsgenome = BSgenome.Hsapiens.UCSC.hg38,
  pwmL = pwmL,
  fixedWidth = 500
)
}

}
