% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{annotateSmallRNA}
\alias{annotateSmallRNA}
\title{Annotate small RNA transcripts to respective categories}
\usage{
annotateSmallRNA(objMOList, anno = "human")
}
\arguments{
\item{objMOList}{A object of class MOList}

\item{anno}{A data frame or a path to a tab-delimited file containing
annotation information for small RNA transcripts. The first
column should be the transcript name of the small RNAs, and
the second column should be the category of the small RNAs.
The category should be one of the following: miRNA, piRNA,
tRNA, circRNA, snRNA, snoRNA. Or, the input can be a simple
string "human" as a shortcut to use the default annotation
provided internally with the package. The package provides
an internal annotation for human small RNAs based on the
GRCm38 genome. The annotation was generated by COMPSRA, using
annotations from the following databases:
1. miRbase
2. piRNABank
3. piRBase
4. GtRNAdb
5. circBase
6. GENCODE
7. piRNACluster
Please see the reference for more details on how the annotation
was generated.}
}
\value{
A MOList object with the annotation information added to the
       annoSncRNA element.
}
\description{
This function annotates small RNA transcripts to their
             respective categories. The categories are defined as follows:
             1. miRNA
             2. piRNA
             3. tRNA
             4. circRNA
             5. snRNA
             6. snoRNA
}
\examples{
# Example 1: Adding user-defined annotation
# Assuming we have an MOList object called objMOList

# Create example annotation
anno <- data.frame(
  transcript = paste0("transcript_", seq_len(120)),
  category = rep(c("miRNA", "piRNA", "tRNA", "circRNA", "snRNA", "snoRNA"),
    each = 20
  )
)

\dontrun{
# Annotate small RNA transcripts
annotateSmallRNA(objMOList, anno)
}

# Example 2: Using the preprocessed annotation provided by the package
\dontrun{
# Annotate small RNA transcripts
annotateSmallRNA(objMOList, anno = "human")

# or simply use the default parameter
annotateSmallRNA(objMOList)
}

}
\references{
Li, J., Kho, A.T., Chase, R.P. et al. COMPSRA: a COMprehensive Platform for
Small RNA-Seq data Analysis. Sci Rep 10, 4552 (2020).
https://doi.org/10.1038/s41598-020-61495-0

Kozomara A, Griffiths-Jones S. miRBase: integrating microRNA annotation and
deep-sequencing data. Nucleic Acids Res. 2011;39 suppl_1:D152–7.

Chan PP, Lowe TM. GtRNAdb 2.0: an expanded database of transfer RNA genes
identified in complete and draft genomes. Nucleic Acids Res. 2016;44:D184–9.

Harrow J, Frankish A, Gonzalez JM, Tapanari E, Diekhans M, Kokocinski F, et
al. GENCODE: The reference human genome annotation for The ENCODE Project.
Genome Res. 2012;22:1760–74.

Glažar P, Papavasileiou P, Rajewsky N. circBase: a database for circular
RNAs. RNA. 2014;20:1666.

Sai lakshmi S, Agrawal S. piRNABank: a web resource on classified and
clustered Piwi-interacting RNAs. Nucleic Acids Res. 2008;36 suppl_1:D173–7.

Zhang P, Si X, Skogerbø G, Wang J, Cui D, Li Y, et al. piRBase: a web
resource assisting piRNA functional study. Database. 2014;2014:1–7.

Rosenkranz D. piRNA cluster database: a web resource for piRNA producing
loci. Nucleic Acids Res. 2016;44:D223–30.
}
