% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEAKTag.R
\name{as.data.frame,PEAKTag-method}
\alias{as.data.frame,PEAKTag-method}
\title{Convert to a data frame}
\usage{
\S4method{as.data.frame}{PEAKTag}(x)
}
\arguments{
\item{x}{A PEAKTag object}
}
\value{
A data frame
}
\description{
Convert to a data frame
}
\examples{
# Use the package's provided example data
data(expMOList)

# Convert the object to a data frame
as.data.frame(expMOList$DEATAC)

}
