% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEAKTag.R
\name{asGRanges}
\alias{asGRanges}
\alias{asGRanges,PEAKTag-method}
\title{Convert to a GRanges object}
\usage{
asGRanges(x)
}
\arguments{
\item{x}{A PEAKTag object}
}
\value{
A GRanges object
}
\description{
Convert to a GRanges object
}
\examples{
# Use the package's provided example data
data(expMOList)

# Convert the object to a GRanges object
\dontrun{
asGRanges(expMOList$DEATAC)
}

}
