% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffOmicsAnalysis.R
\name{countDiffExpr}
\alias{countDiffExpr}
\title{Differential expression analysis of count-based omics data}
\usage{
countDiffExpr(objMOList, omic, batch, program = DESEQ2)
}
\arguments{
\item{objMOList}{A MOList object containing the omics data}

\item{omic}{A character string specifying the omics data to be analyzed
must be one of "RNAseq", "smallRNAseq", and "proteomics"}

\item{batch}{A character vector specifying the batch information for the
omics data, must be the same length as the number of samples in
the omics data, used for batch correction. Can be NULL if no
batch correction is needed.}

\item{program}{A character string specifying the program used for the
analysis, DESeq2 or EdgeR}
}
\value{
An MOList object containing the differential expression analysis
        results. Results are appended to the original MOList object as
        list elements named "DERNAseq", "DEsmallRNAseq", and "DEproteomics"
        for RNAseq, small RNAseq, and protein data, respectively.
\itemize{
\item \code{RNAseq}: A numeric matrix containing the RNAseq data
\item \code{RNAseqSamples}: A list containing the sample names and grouping
                        information for the RNAseq data
\item \code{smallRNAseq}: A numeric matrix containing the smallRNAseq data
\item \code{smallRNAseqSamples}: A list containing the sample names and
                             grouping information for the smallRNAseq data
\item \code{proteomics}: A numeric matrix containing the proteomics data
\item \code{proteomicsSamples}: A list containing the sample names and
                          grouping information for the proteomics data
\item \code{ATACpeaks}: A list containing the ATAC peaks for condition 1 and
                  condition 2
\item \code{DERNAseq}: A DETag object containing the differential
                       expression analysis results for the RNAseq data
\item \code{DEsmallRNAseq}: A DETag object containing the differential
                            expression analysis results for the smallRNAseq
                            data
\item \code{DEproteomics}: A DETag object containing the differential
                           expression analysis results for the proteomics
                           data
}
}
\description{
This function performs differential expression analysis on the
             count-based omics data. The RNAseq, small RNAseq, and protein
             data are supported for differential expression analysis.
             Calling this function again on the same omics data will
             overwrite the previous results.
}
\note{
Preconditions:
\itemize{
\item The input omics data must be a MOList object
\item The input omics data must be filtered by filterGeneCounts
\item The batch information must have the same length as the number of
      samples in the omics data
}
}
\references{
\insertRef{love2014moderated}{IntegraTRN}

\insertRef{robinson2010edger}{IntegraTRN}
}
\keyword{internal}
