% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEOmics.R
\name{countPCA}
\alias{countPCA}
\title{Principle Component Analysis (PCA) for count-based omics data}
\usage{
countPCA(
  matrix,
  groupBy,
  batch = NULL,
  title = NULL,
  col1 = "#440154",
  col2 = "#fde725"
)
}
\arguments{
\item{matrix}{A numeric matrix containing the count-based omics data, must be
raw count expression matrix, not normalized by any method}

\item{groupBy}{A vector of factors for grouping samples}

\item{batch}{A vector of factors for batch effect, default is NULL}

\item{title}{The title for the plot, default is NULL}

\item{col1}{The color for the first group, default is "#440154"}

\item{col2}{The color for the second group, default is "#fde725". If
continuous grouping variable is used, the color gradient will be
used instead of the two colors}
}
\value{
A ggplot object
}
\description{
This function performs PCA analysis for count-based omics data.
             The user can specify the title for the plot.
}
\details{
Raw counts are normalized by RLE method and then transformed by
        variance stabilizing transformation (VST) before PCA analysis.
}
