% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DETag.R
\docType{class}
\name{DETag-class}
\alias{DETag-class}
\title{DETag S4 class}
\description{
This class is used to store the differential expression analysis
             results for the count-based omics data. Although the
             functionality of the class itself can be easily implemented
             by using other data structures, here use specifically designed
             S4 class to allow future extension of the program, e.g.
             providing different methods for differential expression
             analysis while maintaining a consistent interface.
}
\section{Slots}{

\describe{
\item{\code{DEResult}}{A data frame containing the differential expression analysis
results, with each row representing a gene and each column
representing a feature. The exact columns depend on the
method used for the analysis}

\item{\code{method}}{A character string specifying the program used for the analysis,
must be one of the valid METHODS defined}

\item{\code{normalizedCounts}}{A matrix containing the normalized counts for the
genes, with each row representing a gene and each
column representing a sample}
}}

\references{
\insertRef{robinson2010edger}{IntegraTRN}

\insertRef{love2014moderated}{IntegraTRN}
}
