% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DETag.R
\docType{class}
\name{DETag-class}
\alias{DETag-class}
\alias{DETag}
\alias{show,DETag-method}
\title{DETag S4 class}
\usage{
\S4method{show}{DETag}(object)
}
\arguments{
\item{object}{An object of the DETag class}
}
\description{
This class is used to store the differential expression analysis
             results for the count-based omics data. Although the
             functionality of the class itself can be easily implemented
             by using other data structures, here use specifically designed
             S4 class to allow future extension of the program, e.g.
             providing different methods for differential expression
             analysis while maintaining a consistent interface.

This function prints the DETag object
}
\section{Slots}{

\describe{
\item{\code{DEResult}}{A data frame containing the differential expression analysis
results}

\item{\code{method}}{A character string specifying the program used for the analysis,
must be one of the valid METHODS defined}
}}

\examples{
# Create an example data frame
deResult <- data.frame(
  gene = paste0("gene", seq_len(10)),
  logFC = runif(10),
  adj.P.Val = runif(10)
)

# Create an object of the DETag class
deTag <- DETag(deResult, "DESeq2")

# Print the object
print(deTag)

# or simply do
deTag

}
\references{
Robinson MD, McCarthy DJ, Smyth GK. edgeR: a Bioconductor package for
differential expression analysis of digital gene expression data.

Love, M.I., Huber, W., and Anders, S. (2014). Moderated estimation of fold
change and dispersion for RNA-seq data with DESeq2. Genome Biology 15, 1–21.

Advanced R by H. Wickham. Access: https://adv-r.hadley.nz/index.html

Advanced R by H. Wickham. Access: https://adv-r.hadley.nz/index.html
}
