% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DETag.R
\name{DETag}
\alias{DETag}
\title{Constructor for the DETag class}
\usage{
DETag(DEResult, method)
}
\arguments{
\item{DEResult}{A data frame containing the differential expression analysis
results}

\item{method}{A character string specifying the program used for the analysis}
}
\value{
An object of the DETag class
\itemize{
\item \code{DEResult}: A data frame containing the differential expression
                       analysis results
\item \code{method}: A character string specifying the program used for the
                     analysis
}
}
\description{
This function creates an object of the DETag class
}
\examples{
# Create an example data frame
deResult <- data.frame(
  gene = paste0("gene", seq_len(10)),
  logFC = runif(10),
  adj.P.Val = runif(10)
)

# Create an object of the DETag class
deTag <- DETag(deResult, "DESeq2")

# Check the class of the object
class(deTag)

}
\references{
Advanced R by H. Wickham. Access: https://adv-r.hadley.nz/index.html
}
