% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffOmicsAnalysis.R
\name{diffExprDESeq2}
\alias{diffExprDESeq2}
\title{Perform differential expression analysis on the count-based omics data using
DESeq2}
\usage{
diffExprDESeq2(filteredCounts, groupBy, batch = NULL)
}
\arguments{
\item{filteredCounts}{A numeric matrix containing the filtered count-based
omics data}

\item{groupBy}{A vector specifying the grouping information for the omics
data, must be the same length as the number of samples in the
omics data}

\item{batch}{A character vector specifying the batch information for the
omics data, must be the same length as the number of samples in
the omics data, used for batch correction. Can be NULL if no
batch correction is needed.}
}
\value{
A DETag object containing the differential expression analysis
        results, and the method DESeq2
\itemize{
\item \code{DEResult}: A data frame containing the differential expression
                       analysis results
\item \code{method}: The character string "DESeq2"
}
}
\description{
This function performs differential expression analysis on the
             count-based omics data using DESeq2. The RNAseq, small RNAseq,
             and protein data are supported for differential expression
             analysis.
}
\examples{
# Example 1: Differential expression analysis of RNAseq data without batch
#            correction

# Create example count matrix
countMatrix <- matrix(sample(0:100, 1000, replace = TRUE),
  nrow = 100, ncol = 10
)

# Create example grouping information
group <- rep(c("A", "B"), each = 5)

# Perform differential expression analysis
deTag <- diffExprDESeq2(countMatrix, group)

# Check the results
head(exportDE(deTag))

# Example 2: Differential expression analysis of RNAseq data with batch
#            correction

# Create example count matrix
countMatrix <- matrix(sample(0:100, 1000, replace = TRUE),
  nrow = 100, ncol = 10
)

# Create example grouping information
group <- seq_len(10)

# Create example batch information
batch <- rep(c("batch1", "batch2"), each = 5)

# Perform differential expression analysis
deTag <- diffExprDESeq2(
  filteredCounts = countMatrix,
  groupBy = group,
  batch = batch
)

}
\references{
\insertRef{love2014deseq2}{IntegraTRN}
}
\keyword{internal}
