% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffOmicsAnalysis.R
\name{diffOmics}
\alias{diffOmics}
\title{Perform differential analysis on the omics data}
\usage{
diffOmics(
  objMOList,
  rnaseqBatch = NULL,
  smallRnaBatch = NULL,
  proteinBatch = NULL,
  program = "DESeq2"
)
}
\arguments{
\item{objMOList}{A MOList object containing the omics data}

\item{rnaseqBatch}{A character vector specifying the batch information for
the RNA-seq data, must be the same length as the number
of samples in the RNA-seq data, used for batch correction}

\item{smallRnaBatch}{A character vector specifying the batch information for
the small RNA data, must be the same length as the
number of samples in the small RNAseq data, used for
batch correction}

\item{proteinBatch}{A character vector specifying the batch information for
the protein data, must be the same length as the number
of samples in the protein data, used for batch correction}

\item{program}{A character string specifying the program used for the
analysis, DESeq2 or EdgeR}
}
\value{
An MOList object containing the differential analysis results
\itemize{
\item \code{RNAseq}: A numeric matrix containing the RNAseq data
\item \code{RNAseqSamples}: A list containing the sample names and grouping
                        information for the RNAseq data
\item \code{smallRNAseq}: A numeric matrix containing the smallRNAseq data
\item \code{smallRNAseqSamples}: A list containing the sample names and
                             grouping information for the smallRNAseq data
\item \code{proteomics}: A numeric matrix containing the proteomics data
\item \code{proteomicsSamples}: A list containing the sample names and
                          grouping information for the proteomics data
\item \code{ATACpeaks}: A list containing the ATAC peaks for condition 1 and
                  condition 2
\item \code{DERNAseq}: A DETag object containing the differential
                       expression analysis results for the RNAseq data
\item \code{DEsmallRNAseq}: A DETag object containing the differential
                            expression analysis results for the smallRNAseq
                            data
\item \code{DEproteomics}: A DETag object containing the differential
                           expression analysis results for the proteomics
                           data
\item \code{DEATAC}: A DETag object object containing the differential
                     accessible accessible regions for the ATACseq data
}
}
\description{
This function performs data filtering, normalization, batch
             correction, and differential analysis on each of the omics data.
             RNAseq, small RNAseq, and protein data are supported for
             differential expression analysis. The ATACseq data are used
             used to calculate the differential accessible regions.
             While performing the differential analysis, the original count
             data for the RNAseq, small RNAseq, and protein data are
             filtered and subsequently normalized by library size. After
             calling diffOmicsAnalysis, the users can use the S4 methods
             of the MOList class to extract filtered and normalized read
             counts for each of the omics data.
}
