% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DETag.R
\name{exportDE}
\alias{exportDE}
\alias{exportDE,DETag-method}
\title{Export the differential analysis results from a DETag object}
\usage{
exportDE(x, original = FALSE)
}
\arguments{
\item{x}{An object of the DETag class}

\item{original}{A boolean indicating whether to return the original results.
If FALSE, the results will be converted to a data frame with
the following columns: logFC, pvalue, padj
(default: FALSE)}
}
\value{
A data frame containing the differential analysis results, with each
        row representing a gene and each column representing a feature
\itemize{
\item \code{row.names}: A character vector containing the gene names
\item \code{logFC}: A numeric vector containing the log2 fold change values
\item \code{pvalue}: A numeric vector containing the p-values
\item \code{padj}: A numeric vector containing the adjusted p-values
}
}
\description{
This function extracts the differential analysis results from
             the DETag object
}
\examples{
# Use the package-provided example data
data("expMOList")
deTag <- expMOList$DERNAseq

# Example 1: export the package default results
exportDE(deTag)

# Example 2: export the original results as formatted by DESeq2 or edgeR
exportDE(deTag, original = TRUE)

}
\references{
\insertRef{robinson2010edger}{IntegraTRN}

\insertRef{love2014moderated}{IntegraTRN}
}
