% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRNAsmallRNAcoExprEstimate.R
\name{exportMatchResult}
\alias{exportMatchResult}
\title{Export the match result of RNAseq and small RNAseq samples}
\usage{
exportMatchResult(objMOList)
}
\arguments{
\item{objMOList}{An object of class MOList that has been performed matching
using matchSamplesRNAsmallRNA().}
}
\value{
A data frame containing the match result, with the following format:
\itemize{
\item{RNAseq}{The sample names of the RNAseq samples.}
\item{SmallRNAseq}{The sample names of the small RNAseq samples.}
}
}
\description{
This function exports the match result of RNAseq and small
             RNAseq samples.
}
\examples{
# Use the package data
data("expMOList") # data has already been performed matching
# Export the match result
exportMatchResult(expMOList)

}
