% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extractDirectionalGenes}
\alias{extractDirectionalGenes}
\title{Extract a list of names respectively for down and up regulated genes}
\usage{
extractDirectionalGenes(deTag)
}
\arguments{
\item{deTag}{A DETag object}
}
\value{
A list of names respectively for down and up regulated genes
}
\description{
Extract a list of names respectively for down and up regulated genes
}
\keyword{internal}
