% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffOmicsAnalysis.R
\name{filterGeneCounts}
\alias{filterGeneCounts}
\title{Filter the gene counts for the count-based omics data
Level: private}
\usage{
filterGeneCounts(objMOList, omic)
}
\arguments{
\item{objMOList}{A MOList object containing the omics data}

\item{omics}{A character string specifying the omics data to be filtered
must be one of "RNAseq", "smallRNAseq", and "proteomics"}
}
\value{
An MOList object containing the filtered omics data
}
\description{
Filtering is based on the design of the experiment. If the
             samples are only grouped into 2 conditions, then the genes with
             counts in more than 1 counts per million (CPM) in at least 
             min(#samples in condition 1, #samples in condition 2) samples
             are kept. If the samples are grouped by a continuous variable,
             then the genes with counts in more than 1 CPM in at least
             30% of the samples are kept.
}
\examples{
# Create example RNAseq data
rnaseq <- matrix(sample(0:100, 1000, replace = TRUE), nrow = 100, ncol = 10)
rnaseq[1:20, 1:8] <- 0
group <- rep(c("A", "B"), each = 5)

# Create example MOList object
objMOList <- MOList(RNAseq = rnaseq, RNAGroupBy = group)

# Before filtering, check the dimensions of the RNAseq data
dim(getCounts(objMOList, "RNAseq"))

# Filter the RNAseq data
objMOList <- filterGeneCounts(objMOList, "RNAseq")

# After filtering, check the dimensions of the RNAseq data
dim(getCounts(objMOList, "RNAseq"))  # should be lower than the original

}
