% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSampleNames}
\alias{getSampleNames}
\title{Identify the sample names from a count matrix}
\usage{
getSampleNames(countMatrix)
}
\arguments{
\item{countMatrix}{A count matrix with samples as columns and features as
rows}
}
\value{
A vector of sample names
}
\description{
Identify the sample names from a count matrix
}
\examples{
# Create example count matrix
countMatrix <- matrix(sample(0:100, 1000, replace = TRUE),
  nrow = 100, ncol = 10
)

# Get sample names
getSampleNames(countMatrix)

}
\keyword{internal}
