% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jasparVertebratePWM}
\alias{jasparVertebratePWM}
\title{JASPAR curated position weight matrices of transcription factor binding sites
for vertebrates based on database JASPAR CORE 2022}
\format{
A list of 841 position weight matrices of transcription factor
binding sites for vertebrates, as a PWMatrixList object
}
\source{
JASPAR2022 package
}
\usage{
jasparVertebratePWM
}
\description{
A PWMatrixList object containing the position weight matrices of
transcription factor binding sites for vertebrates based on database JASPAR
CORE 2022. This list is extracted from the JASPAR2022 package using the
function \code{getMatrixSet()} from the \code{TFBSTools} package.
}
\examples{
\dontrun{
jasparVertebratePWM
}

}
\references{
Baranasic D (2022). _JASPAR2022: Data package for JASPAR database 
(version 2022)_. doi:10.18129/B9.bioc.JASPAR2022 
<https://doi.org/10.18129/B9.bioc.JASPAR2022>, R package version 0.99.7, 
<https://bioconductor.org/packages/JASPAR2022>.

Tan, G., and Lenhard, B. (2016). TFBSTools: an R/bioconductor package for 
transcription actor binding site analysis. Bioinformatics 32, 1555-1556.
}
\keyword{datasets}
