% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructNetwork.R
\name{loadExtInteractions}
\alias{loadExtInteractions}
\title{Loading external interaction data into the MOList object}
\usage{
loadExtInteractions(
  objMOList,
  upregGenes2miR = NULL,
  downregGenes2miR = NULL,
  upregGenes2TF = NULL,
  downregGenes2TF = NULL
)
}
\arguments{
\item{objMOList}{A MOList object containing the omics data}

\item{upregGenes2miR}{A list containing the upregulated genes and their
                      regulatory miRNAs. Must contain elements "regulator"
                      for the miRNAs and "target" for the upregulated genes.
\itemize{
 \item \code{regulator}: A character vector containing the miRNAs
 \item \code{target}: A character vector containing the upregulated genes
}}

\item{downregGenes2miR}{A list containing the downregulated genes and
their regulatory miRNAs. See above format.}

\item{upregGenes2TF}{A list containing the upregulated genes and their
regulatory transcription factors. See above format.}

\item{downregGenes2TF}{A list containing the downregulated genes and
their regulatory transcription factors. See above
format.}
}
\value{
An object of class MOList, with a element "extInteractions" added,
        which is a list of lists. Each lower level list must follow the
        format of the input data, i.e., must contain elements "regulator"
       and "target". See the examples for details.
\itemize{
\item \code{upregGenes2miR}: A list containing the upregulated genes
                             and their regulatory miRNAs
\item \code{downregGenes2miR}: A list containing the downregulated
                               genes and their regulatory miRNAs
\item \code{upregGenes2TF}: A list containing the upregulated genes
                            and their regulatory transcription factors
\item \code{downregGenes2TF}: A list containing the downregulated
                              genes and their regulatory transcription
                              factors
}
}
\description{
This function loads external interaction data into the MOList
             object. The external data can be curated non-tissue/sample
             specific interaction data, or any interactions that the users
             has determined to encompass a global picture of the regulatory
             interactions of the target genes.
}
\examples{
# Create a sample RNAseq data
RNAseq <- matrix(sample(1:1000, 1000, replace = TRUE), ncol = 10)
colnames(RNAseq) <- paste0("sample_", seq_len(ncol(RNAseq)))
rownames(RNAseq) <- paste0("gene_", seq_len(nrow(RNAseq)))
RNAGroupBy <- rep(c("A", "B"), each = 5)

# Create a myMOList object
myMOList <- MOList(RNAseq = RNAseq, RNAGroupBy = RNAGroupBy)

# Generate some example interaction data
upregGenes2miR <- data.frame(
  ID = paste0("miR_", seq_len(10)),
  Target = paste0("gene_", seq_len(10))
)
downregGenes2miR <- data.frame(
  ID = paste0("miR_", seq(11, 20)),
  Target = paste0("gene_", seq(11, 20))
)
upregGenes2TF <- data.frame(
  ID = paste0("TF_", seq(21, 30)),
  Target = paste0("gene_", seq(21, 30))
)
downregGenes2TF <- data.frame(
  ID = paste0("TF_", seq(31, 40)),
  Target = paste0("gene_", seq(31, 40))
)

# Load the external interaction data into the MOList object
myMOList <- loadExtInteractions(
  myMOList,
  upregGenes2miR = upregGenes2miR,
  downregGenes2miR = downregGenes2miR,
  upregGenes2TF = upregGenes2TF,
  downregGenes2TF = downregGenes2TF
)

}
