% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructNetwork.R
\name{loadExtInteractions}
\alias{loadExtInteractions}
\title{Loading external interaction data into the MOList object}
\usage{
loadExtInteractions(objMOList, miR2Genes = NULL, tf2Genes = NULL)
}
\arguments{
\item{objMOList}{A MOList object containing the omics data}

\item{miR2Genes}{A list containing the target genes and their
                      regulatory miRNAs. Must contain elements "regulator"
                      for the miRNAs and "target" for the upregulated genes.
\itemize{
 \item \code{regulator}: A character vector containing the miRNAs
 \item \code{target}: A character vector containing the upregulated genes
}}

\item{tf2Genes}{A list containing the target genes and their
regulatory transcription factors. See above format.}
}
\value{
An object of class MOList, with a element "extInteractions" added,
        which is a list of lists. Each lower level list must follow the
        format of the input data, i.e., must contain elements "regulator"
       and "target". See the examples for details.
\itemize{
\item \code{miR2Genes}: A list containing the target genes
                             and their regulatory miRNAs
\item \code{tf2Genes}: A list containing the target genes
                            and their regulatory transcription factors
}
}
\description{
This function loads external interaction data into the MOList
             object. The external data can be curated non-tissue/sample
             specific interaction data, or any interactions that the users
             has determined to encompass a global picture of the regulatory
             interactions of the target genes.
}
\examples{
# Create a sample RNAseq data
RNAseq <- matrix(sample(1:1000, 1000, replace = TRUE), ncol = 10)
colnames(RNAseq) <- paste0("sample_", seq_len(ncol(RNAseq)))
rownames(RNAseq) <- paste0("gene_", seq_len(nrow(RNAseq)))
RNAGroupBy <- rep(c("A", "B"), each = 5)

# Create a myMOList object
myMOList <- MOList(RNAseq = RNAseq, RNAGroupBy = RNAGroupBy)

# Generate some example interaction data
miR2Genes <- list(
  regulator = c("miR_1", "miR_2", "miR_3", "miR_4", "miR_5"),
  target = c("gene_1", "gene_2", "gene_3", "gene_4", "gene_5")
)
tf2Genes <- list(
  regulator = c("TF_1", "TF_2", "TF_3", "TF_4", "TF_5"),
  target = c("gene_1", "gene_2", "gene_3", "gene_4", "gene_5")
)
# Load the external interaction data into the MOList object
myMOList <- loadExtInteractions(
  myMOList,
  miR2Genes = miR2Genes,
  tf2Genes = tf2Genes
)

}
