% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matchVecToMatrix}
\alias{matchVecToMatrix}
\title{Match vector length to matrix column length}
\usage{
matchVecToMatrix(vec, mat)
}
\arguments{
\item{vector}{A vector}

\item{matrix}{A matrix}
}
\value{
A boolean indicating whether the vector length matches the matrix
        column length
}
\description{
Match vector length to matrix column length
}
\keyword{internal}
