% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processPeaks.R
\name{mergePeaks}
\alias{mergePeaks}
\title{Merge peaks with overlapping genomic coordinates}
\usage{
mergePeaks(peaks)
}
\arguments{
\item{peaks}{A data frame containing the peak information, in BED format}
}
\value{
A GRanges object containing the merged peaks
}
\description{
Merge peaks with overlapping genomic coordinates
}
\keyword{internal}
