% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mir2geneMultiMiR}
\alias{mir2geneMultiMiR}
\title{Example miRNA - mRNA interactions obtained with miRNA-centric approach}
\format{

}
\source{
miRecords, miRTarBase, and TarBase databases
}
\usage{
mir2geneMultiMiR
}
\description{
A data frame containing the miRNA-target interactions obtained with the
miRNA-centric approach. A list of microRNAs are queried against the
miRecords, miRTarBase, and TarBase databases. 500 interactions are
randomly selected.
}
\examples{
\dontrun{
mir2geneMultiMiR
}

}
\references{
\insertRef{xiao2009mirecords}{IntegraTRN}

\insertRef{huang2020mirtarbase}{IntegraTRN}

\insertRef{karagkouni2018diana}{IntegraTRN}
}
\keyword{datasets}
