% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miR2Genes}
\alias{miR2Genes}
\title{Example miRNA - mRNA interactions}
\format{
A list of 2 vectors of the same length. Each representing a pair of
        miRNA-target interactions.
\describe{\
\item{regulator}{A vector of miRNAs}
\item{target}{A vector of upregulated fetal heart mRNAs}
}
}
\usage{
miR2Genes
}
\description{
A list of miRNA-target interactions on fetal heart differentially expressed
genes. The list contains 2 vectors of the same length. Each representing a
pair of miRNA-target interactions. Curated partially from randomly selected
miRNA-target interactions from miRNet 2.0. Partially simulated.
}
\examples{
\dontrun{
miR2Genes
}

}
\references{
Le Chang, Guangyan Zhou, Othman Soufan, Jianguo Xia, miRNet 2.0:
            network-based visual analytics for miRNA functional analysis and
            systems biology, Nucleic Acids Research, Volume 48, Issue W1, 02
            July 2020, Pages W244–W251, https://doi.org/10.1093/nar/gkaa467
}
\keyword{datasets}
