% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOList.R
\docType{class}
\name{MOList-class}
\alias{MOList-class}
\alias{show,MOList-method}
\alias{setGene2Protein}
\title{Multi-Omics List (MOList) S4 class}
\usage{
\S4method{show}{MOList}(object)

setGene2Protein(x, conversion)
}
\arguments{
\item{object}{An object of the DETag class}

\item{x}{A MOList object containing the omics data}

\item{conversion}{A data frame containing the conversion information between
                  protein and gene names
\itemize{
\item \code{protein}: Name of the protein
\item \code{gene}: Name of the gene
}}
}
\value{
An object of class MOList
}
\description{
This class is used to store the multi-omics data, including
             RNAseq, smallRNAseq, proteomics, and ATAC peaks data.

This function prints the MOList object

This function sets the conversion between protein and gene names
             for the proteomics data. The user can provide a data frame
             containing the conversion information. The data frame must
             contain two columns, one for the protein names and the other
             for the gene names. This must be provided to effectively utilize
             the proteomics data. The users are responsible for the validity
             of the conversion information.
}
\section{Slots}{

\describe{
\item{\code{RNAseq}}{A numeric matrix containing the RNAseq data}

\item{\code{RNAseqSamples}}{A list containing the sample names and grouping
information for the RNAseq data}

\item{\code{smallRNAseq}}{A numeric matrix containing the smallRNAseq data}

\item{\code{smallRNAseqSamples}}{A list containing the sample names and grouping
information for the smallRNAseq data}

\item{\code{proteomics}}{A numeric matrix containing the proteomics data}

\item{\code{proteomicsSamples}}{A list containing the sample names and grouping
information for the proteomics data}

\item{\code{ATACpeaks}}{A list containing the ATAC peaks for condition 1 and
condition 2 that are found to be differentially accessible.
The users should ensure that each BED file used as input
contains the chromosome regions that are found to have
increased accessibility in each condition. The BED regions
can be constructed either by merging the peaks from all
replicates in each condition using a variable-length method
or a fixed-width approach, as indicated in Grandi et al. 2022}

\item{\code{.Data}}{A list containing differential expression analysis data, used
in the same way as a list}
}}

\examples{
\dontrun{
# Assuming myMOList is a MOList object
print(myMOList)
}

\dontrun{
# Assuming myMOList is a MOList object
# Assuming conversion is a data frame containing the conversion information
# between protein and gene names

# Setting the conversion information
setGene2Protein(myMOList, conversion)
}

}
\references{
Advanced R by H. Wickham. Access: https://adv-r.hadley.nz/index.html

Grandi, F.C., Modi, H., Kampman, L. et al. Chromatin accessibility profiling
by ATAC-seq. Nat Protoc 17, 1518–1552 (2022).
https://doi-org/10.1038/s41596-022-00692-9

Advanced R by H. Wickham. Access: https://adv-r.hadley.nz/index.html
}
