% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOList.R
\docType{class}
\name{MOList-class}
\alias{MOList-class}
\title{Multi-Omics List (MOList) S4 class}
\description{
This class is used to store the multi-omics data, including
             RNAseq, smallRNAseq, proteomics, and ATAC peaks data.
}
\section{Slots}{

\describe{
\item{\code{RNAseq}}{A numeric matrix containing the RNAseq data}

\item{\code{RNAseqSamples}}{A list containing the sample names and grouping
information for the RNAseq data}

\item{\code{smallRNAseq}}{A numeric matrix containing the smallRNAseq data}

\item{\code{smallRNAseqSamples}}{A list containing the sample names and grouping
information for the smallRNAseq data}

\item{\code{proteomics}}{A numeric matrix containing the proteomics data}

\item{\code{proteomicsSamples}}{A list containing the sample names and grouping
information for the proteomics data}

\item{\code{ATACpeaks}}{A list containing the ATAC peaks for condition 1 and
condition 2 that are found to be differentially accessible.
The users should ensure that each BED file used as input
contains the chromosome regions that are found to have
increased accessibility in each condition. The BED regions
can be constructed either by merging the peaks from all
replicates in each condition using a variable-length method
or a fixed-width approach, as indicated in Grandi et al. 2022}

\item{\code{.Data}}{A list containing differential expression analysis data, used
in the same way as a list}
}}

\references{
\insertRef{grandi2022chromatin}{IntegraTRN}
}
