% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEAKTag.R
\docType{class}
\name{PEAKTag-class}
\alias{PEAKTag-class}
\alias{show,PEAKTag-method}
\alias{asGRanges}
\title{PEAKTag S4 class for storing peak information}
\usage{
\S4method{show}{PEAKTag}(object)

asGRanges(x)
}
\arguments{
\item{object}{An object of the DETag class}

\item{x}{A PEAKTag object}
}
\value{
A GRanges object
}
\description{
The PEAKTag S4 class inherits the DETag class and is a more
             specialized class for storing peak information. It contains
             all slots of the DETag class, with additional slots added as a
             more convenient way for users to access the peak information.

The print method for the PEAKTag class.
}
\section{Slots}{

\describe{
\item{\code{DEResult}}{A data frame containing the annotated peak information
\itemize{
 \item \code{chr}: Chromosome
\item \code{start}: Start coordinate
\item \code{end}: End coordinate
\item \code{width}: Width of the peak
\item \code{strand}: Strand, always \code{"*"} for ATACseq
\item \code{Condition}: Condition of the peak, "-" for peaks from condition 1
                        and "+" for peaks from condition 2
\item \code{annotation}: Annotation of the peak, the class of the peak
\item \code{ENSEMBL}: ENSEMBL gene ID, based on the peak annotation
\item \code{SYMBOL}: Gene symbol, based on the peak annotation
\item \code{GENENAME}: Gene name, based on the peak annotation
}}

\item{\code{method}}{A character string indicating the method used to generate the
peak set, defined similarly to the DETag class. In this case,
must be "GRangesATAC"}

\item{\code{normalizedCounts}}{A matrix containing the normalized counts for the
genes. Must be NULL for this class. However, this
slot is kept for compatibility with the DETag class
and allows future expansion of the package to support
differential peak calling using sequence coverage
information.}

\item{\code{annotatedPeaks}}{A csAnno object containing the annotated peaks, as
defined in the ChIPseeker package}

\item{\code{TxDB}}{A TxDb object containing the transcript database used for the
annotation}

\item{\code{annoDB}}{A string indicating the annotation database used for the
annotation}

\item{\code{.Data}}{A list containing additional peak information}
}}

\examples{
# Assuming that the object "peakTag" is a PEAKTag object
\dontrun{
print(peakTag)
}

# Or simply type the object name
\dontrun{
peakTag
}

# Assuming that the object "peakTag" is a PEAKTag object
\dontrun{
asGRanges(peakTag)
}

}
