% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEAKTag.R
\name{PEAKTag}
\alias{PEAKTag}
\title{Constructor for the PEAKTag class}
\usage{
PEAKTag(objDETag, annotatedPeaks = NULL, TxDB = NULL, annoDB = NULL)
}
\arguments{
\item{objDETag}{A DETag object}

\item{annotatedPeaks}{A csAnno object containing the annotated peaks, as
defined in the ChIPseeker package}

\item{TxDB}{A TxDb object containing the transcript database used for the
annotation}

\item{annoDB}{A string indicating the annotation database used for the
annotation}
}
\value{
A PEAKTag object
}
\description{
This function constructs a PEAKTag object from a DETag object
             and the annotated peaks, as defined in the ChIPseeker package.
}
