% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEOmics.R
\name{plotATACAnno}
\alias{plotATACAnno}
\title{Plotting the annotation Pie chart of differential accessible regions}
\usage{
plotATACAnno(objMOList)
}
\arguments{
\item{objMOList}{An MOList object containing the differential accessible
regions}
}
\value{
A ggplot object
}
\description{
This function generates a pie chart for the annotation of
             differential accessible regions based on the ATACseq data.
             This is a wrapper function for the plotAnnoPie function in the
             ChIPseeker package.
}
\examples{
# Suppose that we have an MOList object called objMOList, which contains the
# differential accessible regions with annotation

# Plotting the annotation pie chart
\dontrun{
plotATACAnno(objMOList)
}

}
