% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEOmics.R
\name{plotATACCoverage}
\alias{plotATACCoverage}
\title{Peak coverage plot for differential accessible regions}
\usage{
plotATACCoverage(objMOList, title = "ATAC Peaks over Chromosomes")
}
\arguments{
\item{objMOList}{An MOList object containing the differential accessible
regions}

\item{title}{The title for the plot, default is "ATAC Peaks over Chromosomes"}
}
\value{
A ggplot object
}
\description{
This function generates a peak coverage plot for differential
             accessible regions based on the ATACseq data. This is a wrapper
             function for the covplot function in the ChIPseeker package.
}
\examples{
# Use the package-provided example data
data(expMOList)

# Plotting the coverage plot
\dontrun{
plotATACCoverage(expMOList)
}

}
\references{
\insertRef{yu2015chipseeker}{IntegraTRN}

\insertRef{lawrence2013software}{IntegraTRN}
}
