% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEOmics.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Generate a base volcano plot}
\usage{
plotVolcano(deg, log2FC, adjP, upColor, downColor, title)
}
\arguments{
\item{deg}{A data frame containing the differential expression analysis
results, must follow the described format in DETag-class}

\item{log2FC}{The cutoff for log2 fold change.}

\item{adjP}{The cutoff for adjusted p-value.}

\item{upColor}{The color for up-regulated genes.}

\item{downColor}{The color for down-regulated genes.}

\item{title}{The title for the plot.}
}
\value{
A ggplot object
}
\description{
This function generates a base volcano plot for visualizing
            differentially expressed omics data based on log2 fold change
            and adjusted p-value.
}
\keyword{internal}
