% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEOmics.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Generate a base volcano plot}
\usage{
plotVolcano(deg, log2FC = 1, adjP = 0.05, title = NULL)
}
\arguments{
\item{deg}{A data frame containing the differential expression analysis
results, must follow the described format in DETag-class}

\item{log2FC}{The cutoff for log2 fold change.}

\item{adjP}{The cutoff for adjusted p-value.}

\item{title}{The title for the plot.}
}
\value{
A ggplot object
}
\description{
This function generates a base volcano plot for visualizing
            differentially expressed omics data based on log2 fold change
            and adjusted p-value.
}
\references{
\insertRef{villanueva2019ggplot2}{IntegraTRN}

\insertRef{dplyr}{IntegraTRN}
}
\keyword{internal}
