% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructNetwork.R
\name{print.OMICutoffs}
\alias{print.OMICutoffs}
\title{Print method for the OMICutoffs object}
\usage{
\method{print}{OMICutoffs}(x, ...)
}
\arguments{
\item{x}{An OMICutoffs object}

\item{...}{Other arguments passed to the print function}
}
\value{
The cutoffs for the omics data
}
\description{
This function prints the cutoffs for the omics data.
}
\examples{
omiCutoffs <- setOmicCutoffs()

# The output will be:
# RNAseq adjusted p-value cutoff: 0.05
# RNAseq log fold change cutoff: 1
# RNAseq selecting top 100\% of DE genes
# smallRNAseq adjusted p-value cutoff: 0.05
# smallRNAseq log fold change cutoff: 1
# smallRNAseq selecting top 100\% of DE genes
# Proteomics adjusted p-value cutoff: 0.05
# Proteomics log fold change cutoff: 1

}
