% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOList.R
\name{proteomics,MOList-method}
\alias{proteomics,MOList-method}
\alias{proteomics}
\title{Getter for the proteomics data from the MOList object}
\usage{
\S4method{proteomics}{MOList}(x)
}
\arguments{
\item{x}{An object of class MOList for retrieving omics data}
}
\value{
A numeric matrix containing the proteomics data
}
\description{
Getter for the proteomics data from the MOList object
}
\examples{
# Generating some example data
# Note that the exact value could differ based on the random seed
RNAseq <- matrix(sample(1:100, 100, replace = TRUE), ncol = 10)
RNAGroupBy <- rep(c("A", "B"), each = 5)
proteomics <- matrix(sample(1:100, 30, replace = TRUE), ncol = 6)
proteomicsGroupBy <- rep(c("A", "B"), each = 3)

# Constructing the MOList object
objMOList <- MOList(RNAseq = RNAseq,
                    RNAGroupBy = RNAGroupBy,
                    proteomics = proteomics,
                    proteomicsGroupBy = proteomicsGroupBy)

# Retrieving the proteomics data
dataProteomics <- proteomics(objMOList)

}
