% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOList.R
\name{RNAseq,MOList-method}
\alias{RNAseq,MOList-method}
\alias{RNAseq}
\title{Getter for the RNAseq data from the MOList object}
\usage{
\S4method{RNAseq}{MOList}(x)
}
\arguments{
\item{x}{An object of class MOList for retrieving omics data}
}
\value{
A numeric matrix containing the RNAseq data
}
\description{
Getter for the RNAseq data from the MOList object
}
\examples{
# Generating some example data
# Note that the exact value could differ based on the random seed
RNAseq <- matrix(sample(1:100, 100, replace = TRUE), ncol = 10)
RNAGroupBy <- rep(c("A", "B"), each = 5)

# Constructing the MOList object
objMOList <- MOList(RNAseq = RNAseq, RNAGroupBy = RNAGroupBy)

# Retrieving the RNAseq data
dataRNAseq <- RNAseq(objMOList)

}
