% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RNAseq_heart}
\alias{RNAseq_heart}
\title{Gene expression of fetal heart tissues}
\format{
A matrix with 13042 rows (genes) and 53 columns (samples)
\describe{
 \item{sample_1, sample_2, sample_3, sample_4, sample_5, sample_6, sample_7,
 sample_8, sample_9, sample_10, sample_11, sample_12, sample_13, sample_14,
 sample_15, sample_16, sample_17, sample_18, sample_19, sample_20, sample_21,
 sample_22, sample_23, sample_24, sample_25, sample_26, sample_27, sample_28,
 sample_29, sample_30, sample_31, sample_32, sample_33, sample_34, sample_35,
 sample_36, sample_37, sample_38, sample_39, sample_40, sample_41, sample_42,
 sample_43, sample_44, sample_45, sample_46, sample_47, sample_48, sample_49,
 sample_50, sample_51, sample_52, sample_53}{Gene expression values for each
 sample}
}
}
\source{
Paul Delgado Olguin Lab, The Hospital for Sick Children, Toronto,
ON, Canada
}
\usage{
RNAseq_heart
}
\description{
An RNAseq experiment conducted on fetal ventricular heart tissues from 53
individuals
}
\examples{
\dontrun{
RNAseq_heart
}

}
\references{
Ahmed, A., Liang, M., Chi, L., Zhou, Y. Q., Sled, J. G., Wilson, M. D., &
Delgado-Olguín, P. (2021). Maternal obesity persistently alters cardiac
progenitor gene expression and programs adult-onset heart disease
susceptibility. Molecular Metabolism, 43.
https://doi.org/10.1016/j.molmet.2020.101116
}
\keyword{datasets}
