% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DETag.R
\name{show,DETag-method}
\alias{show,DETag-method}
\alias{print,DETag-method}
\title{Print DETag}
\usage{
\S4method{show}{DETag}(object)
}
\arguments{
\item{object}{An object of the DETag class}
}
\description{
This function prints the DETag object
}
\examples{
# Create an example data frame
deResult <- data.frame(
  gene = paste0("gene", seq_len(10)),
  logFC = runif(10),
  adj.P.Val = runif(10)
)

# Create an object of the DETag class
deTag <- DETag(deResult, "DESeq2")

# Print the object
print(deTag)

# or simply do
deTag

}
