% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOList.R
\name{smallRNAseq,MOList-method}
\alias{smallRNAseq,MOList-method}
\alias{smallRNAseq}
\title{Getter for the smallRNAseq data from the MOList object}
\usage{
\S4method{smallRNAseq}{MOList}(x)
}
\arguments{
\item{x}{An object of class MOList for retrieving omics data}
}
\value{
A numeric matrix containing the smallRNAseq data
}
\description{
Getter for the smallRNAseq data from the MOList object
}
\examples{
\dontrun{
# Using the example MOList object
dataSmallRNAseq <- smallRNAseq(myMOList)
}

}
