% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tf2Genes}
\alias{tf2Genes}
\title{Example TF - mRNA interactions}
\format{
A list of 2 vectors of the same length. Each representing a pair of
       TF-target interactions.
\describe{\
\item{regulator}{A vector of TFs}
\item{target}{A vector of upregulated fetal heart mRNAs}
}
}
\usage{
tf2Genes
}
\description{
A list of TF-target interactions on fetal heart differentially expressed
genes. The list contains 2 vectors of the same length. Each representing a
pair of TF-target interactions. Curated partially from randomly selected
TF-target interactions from TRRUST v2 using miRNet 2.0. Partially simulated
using simulated ATACseq peaks.
}
\examples{
\dontrun{
tf2Genes
}

}
\references{
Le Chang, Guangyan Zhou, Othman Soufan, Jianguo Xia, miRNet 2.0:
            network-based visual analytics for miRNA functional analysis and
            systems biology, Nucleic Acids Research, Volume 48, Issue W1, 02
            July 2020, Pages W244–W251, https://doi.org/10.1093/nar/gkaa467
}
\keyword{datasets}
