% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOPTag.R
\docType{class}
\name{TOPTag-class}
\alias{TOPTag-class}
\alias{show,TOPTag-method}
\title{TOPTag S4 class for storing top differential genes}
\usage{
\S4method{show}{TOPTag}(object)
}
\arguments{
\item{object}{An object of the DETag class}
}
\value{
An object of the DETag class
\itemize{
\item \code{DEResult}: A data frame containing the differential expression
                       analysis results
\item \code{method}: A character string specifying the program used for the
                     analysis
}
}
\description{
The TOPTag S4 class inherits the DETag class and is a more
             specific class for storing top differential genes. It contains
             all slots of the DETag class, with stringent filtering on the
             genes involved. Additional slots are added as a more convenient
             way for users to access the top differential genes.

The print method for the TOPTag class.
}
\section{Slots}{

\describe{
\item{\code{DEResult}}{A data frame containing the differential expression results
               for only the top differential genes. The columns resemble the
               predefined format in the DETag class. An additional column
               "rank" is added to indicate the rank of the gene in the
               differential expression results.
\itemize{
 \item {logFC}{The log fold change of the gene}
 \item {pvalue}{The p-value of the gene}
 \item {padj}{The adjusted p-value of the gene}
 \item {piValue}{The pi-value of the gene, calculated by the formula
                -log10(padj) * sign(logFC)}
 \item {rank}{The rank of the gene in the differential expression results.
              Positive for up-regulated genes and negative for down-
              regulated genes. A rank of 1 indicates the most up-regulated
              gene and a rank of -1 indicates the most down-regulated gene.}
}}

\item{\code{method}}{A character string indicating the method used to calculate the
differential expression results.}

\item{\code{normalizedCounts}}{A matrix containing the normalized counts for
only the top differential genes, consistent with the
the DEResult slot.}

\item{\code{logFCCutoff}}{A numeric value indicating the absolute log2 fold change
cutoff used to filter the differential expression results.}

\item{\code{pCutoff}}{A numeric value indicating the adjusted p-value cutoff used to
filter the differential expression results.}

\item{\code{topGenes}}{A numeric value indicating either the fraction (0-1) of top
differential genes or the number (1-Inf) of top differential
genes. If the number specified is greater than the number of
DE genes based on logFCCutoff and pCutoff, then topGenes will
be set to the number of DE genes.}
}}

\examples{
# Assuming that the object "topTag" is a TOPTag object
\dontrun{
print(topTag)
}

# Or simply type the object name
\dontrun{
topTag
}

}
\references{
Robinson MD, McCarthy DJ, Smyth GK. edgeR: a Bioconductor package for
differential expression analysis of digital gene expression data.

Love, M.I., Huber, W., and Anders, S. (2014). Moderated estimation of fold
change and dispersion for RNA-seq data with DESeq2. Genome Biology 15, 1–21.

Advanced R by H. Wickham. Access: https://adv-r.hadley.nz/index.html

Reimand J, Isserlin R, Voisin V, et al. Pathway enrichment analysis and
visualization of omics data using g:Profiler, GSEA, Cytoscape and
EnrichmentMap. Nat Protoc. 2019;14(2):482-517. doi:10.1038/s41596-018-0103-9.
}
