% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOPTag.R
\docType{class}
\name{TOPTag-class}
\alias{TOPTag-class}
\alias{show,TOPTag-method}
\alias{filterGenes}
\alias{exportDE,TOPTag-method}
\title{TOPTag S4 class for storing top differential genes}
\usage{
\S4method{show}{TOPTag}(object)

filterGenes(object, names)

\S4method{exportDE}{TOPTag}(x, original = FALSE)
}
\arguments{
\item{object}{A TOPTag object.}

\item{names}{A character vector containing the names of the genes to be
filtered.}

\item{x}{An object of the DETag class}

\item{original}{A boolean indicating whether to return the original results.
If FALSE, the results will be converted to a data frame with
the following columns: logFC, pvalue, padj
(default: FALSE)}
}
\value{
A TOPTag object.

A data frame containing the differential analysis results.
}
\description{
The TOPTag S4 class inherits the DETag class and is a more
             specific class for storing top differential genes. It contains
             all slots of the DETag class, with stringent filtering on the
             genes involved. Additional slots are added as a more convenient
             way for users to access the top differential genes.

The print method for the TOPTag class.

This function filters the top differential genes by a names
             vector

This function exports the differential analysis results from a
             TopTag object. Overwrites the exportDE function in the DETag
             class due to different nomenclature of the columns.
}
\section{Slots}{

\describe{
\item{\code{DEResult}}{A data frame containing the differential expression results
               for only the top differential genes. The columns resemble the
               predefined format in the DETag class. An additional column
               "rank" is added to indicate the rank of the gene in the
               differential expression results.
\itemize{
 \item {logFC}{The log fold change of the gene}
 \item {pvalue}{The p-value of the gene}
 \item {padj}{The adjusted p-value of the gene}
 \item {piValue}{The pi-value of the gene, calculated by the formula
                -log10(padj) * sign(logFC)}
 \item {rank}{The rank of the gene in the differential expression results.
              Positive for up-regulated genes and negative for down-
              regulated genes. A rank of 1 indicates the most up-regulated
              gene and a rank of -1 indicates the most down-regulated gene.}
}}

\item{\code{method}}{A character string indicating the method used to calculate the
differential expression results.}

\item{\code{normalizedCounts}}{A matrix containing the normalized counts for
only the top differential genes, consistent with the
the DEResult slot.}

\item{\code{logFCCutoff}}{A numeric value indicating the absolute log2 fold change
cutoff used to filter the differential expression results.}

\item{\code{pCutoff}}{A numeric value indicating the adjusted p-value cutoff used to
filter the differential expression results.}

\item{\code{topGenes}}{A numeric value indicating either the fraction (0-1) of top
differential genes or the number (1-Inf) of top differential
genes. If the number specified is greater than the number of
DE genes based on logFCCutoff and pCutoff, then topGenes will
be set to the number of DE genes.}
}}

\examples{
# Use the package provided example data
data("expMOList")

# Extract an DETag object from RNA-seq data
deTag <- expMOList$DERNAseq

# Create a TOPTag object
topTag <- TOPTag(deTag, topGenes = 0.2)

# Print the TOPTag object
print(topTag)

# or simply type the object name
topTag

# Use the package provided example data
data("expMOList")

# Extract an DETag object from RNA-seq data
deTag <- expMOList$DERNAseq

# Create a TOPTag object
topTag <- TOPTag(deTag, logFCCutoff = 0, topGenes = 0.2)

# Filter the top differential genes by a names vector
topTag <- filterGenes(topTag, c("B2M", "VSIR", "HAND2-AS1"))

# Use the package provided example data
data("expMOList")

# Extract an DETag object from RNA-seq data
deTag <- expMOList$DERNAseq

# Create a TOPTag object
topTag <- TOPTag(deTag, logFCCutoff = 0, topGenes = 0.2)

# Export the differential analysis results
exportDE(topTag)

}
\references{
\insertRef{love2014moderated}{IntegraTRN}

\insertRef{robinson2010edger}{IntegraTRN}

\insertRef{reimand2019pathway}{IntegraTRN}
}
