% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOPTag.R
\name{TOPTag}
\alias{TOPTag}
\title{Constructor for TOPTag class}
\usage{
TOPTag(
  object,
  logFCCutoff = 1,
  pCutoff = 0.05,
  topGenes = 0.1,
  direction = "both"
)
}
\arguments{
\item{object}{A DETag object.}

\item{logFCCutoff}{A numeric value indicating the absolute log2 fold change
cutoff used to filter the differential expression results.}

\item{pCutoff}{A numeric value indicating the adjusted p-value cutoff used to
filter the differential expression results.}

\item{topGenes}{A numeric value indicating either the fraction (0-1) of top
differential genes or the number (1-Inf) of top differential
genes. If the number specified is greater than the number of
DE genes based on logFCCutoff and pCutoff, then topGenes will
be set to the number of DE genes. Default to select the top
10% of differential genes.}

\item{direction}{A character string indicating the direction of the
differential expression. Default to "both". Other options
include "up" and "down".}
}
\value{
A TOPTag object.
}
\description{
The constructor for the TOPTag class.
}
\examples{
# Assuming that the object "deTag" is a DETag object

# Example 1: Select the top 20\% of differential genes with default cutoffs

\dontrun{
topTag <- TOPTag(deTag, topGenes = 0.2)
}

# Example 2: Select the top 100 differential genes with default cutoffs

\dontrun{
topTag <- TOPTag(deTag, topGenes = 100)
}

# Example 3: Select the top 20\% of differential genes with custom cutoffs

\dontrun{
topTag <- TOPTag(deTag, logFCCutoff = 2, pCutoff = 0.01, topGenes = 0.2)
}

}
