% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRNet.R
\docType{class}
\name{TRNet-class}
\alias{TRNet-class}
\alias{generatePlot}
\title{TRNet S4 class of transcriptional regulatory network}
\usage{
generatePlot(trn)
}
\arguments{
\item{trn}{A TRNet object}
}
\value{
A TRNet object with the igraph object in the network slot
}
\description{
The TRNet S4 class is a class for storing transcriptional
             regulatory network. It internally utilizes the igraph package
             to store the network, with additional parameters to store
             information about the network.

This function generates an igraph object from the TRNet network
             metadata
}
\section{Slots}{

\describe{
\item{\code{network}}{An igraph object containing the transcriptional regulatory
network. The nodes of the network are the genes and the edges
are the regulatory relationships between the genes. This is a
named, directed, and weighted graph.}

\item{\code{TRNmetadata}}{A data frame containing all interaction data for the
transcriptional regulatory network in adjacency list
format. The columns minimally includes "regulator",
"target", and "regulatorType".}

\item{\code{predicted}}{A logical value indicating whether the transcriptional
regulatory network contains inferred interactions.}

\item{\code{omics}}{A character string indicating the omics data used to construct
the transcriptional regulatory network.}

\item{\code{.Data}}{A list containing additional network information}
}}

\references{
\insertRef{csardi2006igraph}{IntegraTRN}
}
