% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DETag.R
\name{validateDETagSlots}
\alias{validateDETagSlots}
\title{Validator of the DETag class slots}
\usage{
validateDETagSlots(DEResult, method, normalizedCounts = NULL)
}
\arguments{
\item{DEResult}{A data frame containing the differential expression analysis
results as the outputs from respective programs, with each
row representing a gene and each column representing a
feature. The exact columns depend on the method used for the
analysis}

\item{method}{A character string specifying the program used for the
analysis, must be one of the valid METHODS defined}
}
\description{
This function validates the DETag class
}
\keyword{internal}
