% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOList.R
\name{validateMOList}
\alias{validateMOList}
\title{Validator of the MOList S4 object}
\usage{
validateMOList(objMOList)
}
\arguments{
\item{objMOList}{An object of class MOList for appending/exchanging omics
data}
}
\description{
Validator of the MOList S4 object
}
\keyword{internal}
