% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOList.R
\name{validateOmics}
\alias{validateOmics}
\title{Validate omics raw data with sample information}
\usage{
validateOmics(dataMatrix, sampleInfo)
}
\arguments{
\item{dataMatrix}{A numeric matrix containing the omics data}

\item{sampleInfo}{A list containing the sample names and grouping information
for the omics data}
}
\description{
Validate omics raw data with sample information
}
\keyword{internal}
